+++ bin/informational/overdrive-advantage-accounts
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
 
-from palace.manager.api.overdrive import GenerateOverdriveAdvantageAccountList
+from palace.manager.api.overdrive.script import GenerateOverdriveAdvantageAccountList
 
 GenerateOverdriveAdvantageAccountList().run()
