+++ bin/informational/overdrive-raw-circulation
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import json
 
-from palace.manager.api.overdrive import OverdriveAPI
+from palace.manager.api.overdrive.api import OverdriveAPI
 from palace.manager.scripts.input import IdentifierInputScript
 from palace.manager.sqlalchemy.model.collection import Collection
