+++ bin/overdrive_monitor_recent
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 """Monitor the Overdrive collections by going through the recently changed list."""
 
-
-from palace.manager.api.overdrive import RecentOverdriveCollectionMonitor
+from palace.manager.api.overdrive.monitor import RecentOverdriveCollectionMonitor
 from palace.manager.scripts.monitor import RunCollectionMonitorScript
 
 RunCollectionMonitorScript(RecentOverdriveCollectionMonitor).run()
