+++ bin/overdrive_monitor_search
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 """Monitor the Overdrive collections using Overdrive's search feature"""
 
-
-from palace.manager.api.overdrive import OverdriveCirculationMonitor
+from palace.manager.api.overdrive.monitor import OverdriveCirculationMonitor
 from palace.manager.scripts.monitor import RunCollectionMonitorScript
 
 RunCollectionMonitorScript(OverdriveCirculationMonitor).run()
