+++ bin/overdrive_new_titles
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 """Look for new titles added to Overdrive collections which slipped through the cracks."""
 
-
-from palace.manager.api.overdrive import NewTitlesOverdriveCollectionMonitor
+from palace.manager.api.overdrive.monitor import NewTitlesOverdriveCollectionMonitor
 from palace.manager.scripts.monitor import RunCollectionMonitorScript
 
 RunCollectionMonitorScript(NewTitlesOverdriveCollectionMonitor).run()
