+++ src/palace/manager/api/overdrive/api.py
@@ -1,41 +1,21 @@
 from __future__ import annotations
 
-import argparse
-import csv
 import datetime
 import json
-import logging
 import re
-import traceback
 import urllib.parse
 from collections.abc import Callable, Generator, Iterable, Mapping
 from json import JSONDecodeError
 from threading import RLock
 from typing import Any, NamedTuple, cast
-from urllib.parse import quote, urlsplit, urlunsplit
 
-import dateutil
 import flask
-import isbnlib
-from dependency_injector.wiring import Provide, inject
-from flask_babel import lazy_gettext as _
 from requests import Response
 from requests.structures import CaseInsensitiveDict
-from sqlalchemy import select
 from sqlalchemy.orm import Session
-from sqlalchemy.orm.exc import ObjectDeletedError, StaleDataError
-from tenacity import (
-    retry,
-    retry_if_exception_type,
-    stop_after_attempt,
-    wait_exponential,
-)
-from typing_extensions import Self
 
 from palace.manager.api.circulation import (
     BaseCirculationAPI,
-    BaseCirculationApiSettings,
-    BaseCirculationEbookLoanSettings,
     CirculationInternalFormatsMixin,
     DeliveryMechanismInfo,
     FetchFulfillment,
@@ -61,174 +41,43 @@ from palace.manager.api.circulation_exceptions import (
     PatronHoldLimitReached,
     PatronLoanLimitReached,
 )
+from palace.manager.api.overdrive.advantage import OverdriveAdvantageAccount
+from palace.manager.api.overdrive.constants import (
+    OVERDRIVE_LABEL,
+    OVERDRIVE_MAIN_ACCOUNT_ID,
+    OverdriveConstants,
+)
+from palace.manager.api.overdrive.coverage import OverdriveBibliographicCoverageProvider
+from palace.manager.api.overdrive.fulfillment import OverdriveManifestFulfillment
+from palace.manager.api.overdrive.representation import OverdriveRepresentationExtractor
+from palace.manager.api.overdrive.settings import (
+    OverdriveChildSettings,
+    OverdriveLibrarySettings,
+    OverdriveSettings,
+)
+from palace.manager.api.overdrive.util import _make_link_safe
 from palace.manager.api.selftest import HasCollectionSelfTests, SelfTestResult
 from palace.manager.core.config import CannotLoadConfiguration, Configuration
-from palace.manager.core.coverage import BibliographicCoverageProvider, CoverageFailure
-from palace.manager.core.exceptions import (
-    BasePalaceException,
-    IntegrationException,
-    PalaceValueError,
-)
-from palace.manager.core.metadata_layer import (
-    CirculationData,
-    ContributorData,
-    FormatData,
-    IdentifierData,
-    LinkData,
-    MeasurementData,
-    Metadata,
-    ReplacementPolicy,
-    SubjectData,
-    TimestampData,
-)
-from palace.manager.core.monitor import (
-    CollectionMonitor,
-    IdentifierSweepMonitor,
-    TimelineMonitor,
-)
-from palace.manager.integration.base import (
-    HasChildIntegrationConfiguration,
-    integration_settings_update,
-)
-from palace.manager.integration.configuration.connection import ConnectionSetting
-from palace.manager.integration.goals import Goals
-from palace.manager.integration.settings import (
-    BaseSettings,
-    ConfigurationFormItem,
-    ConfigurationFormItemType,
-    FormField,
-)
-from palace.manager.scripts.base import Script
-from palace.manager.scripts.input import InputScript
-from palace.manager.service.analytics.analytics import Analytics
-from palace.manager.service.container import Services
-from palace.manager.sqlalchemy.constants import DataSourceConstants, MediaTypes
-from palace.manager.sqlalchemy.model.classification import Classification, Subject
+from palace.manager.core.exceptions import BasePalaceException, IntegrationException
+from palace.manager.core.metadata_layer import ReplacementPolicy
+from palace.manager.integration.base import HasChildIntegrationConfiguration
+from palace.manager.sqlalchemy.constants import MediaTypes
 from palace.manager.sqlalchemy.model.collection import Collection
-from palace.manager.sqlalchemy.model.contributor import Contributor
 from palace.manager.sqlalchemy.model.credential import Credential
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.identifier import Identifier
-from palace.manager.sqlalchemy.model.integration import IntegrationConfiguration
 from palace.manager.sqlalchemy.model.library import Library
 from palace.manager.sqlalchemy.model.licensing import (
     DeliveryMechanism,
     LicensePool,
     LicensePoolDeliveryMechanism,
 )
-from palace.manager.sqlalchemy.model.measurement import Measurement
 from palace.manager.sqlalchemy.model.patron import Loan, Patron
-from palace.manager.sqlalchemy.model.resource import Hyperlink, Representation
+from palace.manager.sqlalchemy.model.resource import Representation
 from palace.manager.util import base64
 from palace.manager.util.datetime_helpers import strptime_utc, utc_now
 from palace.manager.util.http import HTTP, BadResponseException
-from palace.manager.util.log import LoggerMixin
-
-
-class OverdriveConstants:
-    PRODUCTION_SERVERS = "production"
-    TESTING_SERVERS = "testing"
-
-    # The formats we care about.
-    FORMATS = "ebook-epub-open,ebook-epub-adobe,ebook-pdf-adobe,ebook-pdf-open,audiobook-overdrive".split(
-        ","
-    )
-
-    # These are not real Overdrive formats; we use them internally so
-    # we can distinguish between (e.g.) using "audiobook-overdrive"
-    # to get into Overdrive Read, and using it to get a link to a
-    # manifest file.
-    MANIFEST_INTERNAL_FORMATS = {
-        "audiobook-overdrive-manifest",
-        "ebook-overdrive-manifest",
-    }
-
-    # These formats can be delivered either as manifest files or as
-    # links to websites that stream the content.
-    STREAMING_FORMATS = [
-        "ebook-overdrive",
-        "audiobook-overdrive",
-    ]
-
-    # When associating an Overdrive account with a library, it's
-    # necessary to also specify an "ILS name" obtained from
-    # Overdrive. Components that don't authenticate patrons (such as
-    # the metadata wrangler) don't need to set this value.
-    ILS_NAME_KEY = "ils_name"
-    ILS_NAME_DEFAULT = "default"
-
-
-class OverdriveSettings(ConnectionSetting, BaseCirculationApiSettings):
-    """The basic Overdrive configuration"""
-
-    external_account_id: str | None = FormField(
-        form=ConfigurationFormItem(
-            label=_("Library ID"),
-            type=ConfigurationFormItemType.TEXT,
-            description="The library identifier.",
-            required=True,
-        ),
-    )
-    overdrive_website_id: str = FormField(
-        form=ConfigurationFormItem(
-            label=_("Website ID"),
-            type=ConfigurationFormItemType.TEXT,
-            description="The web site identifier.",
-            required=True,
-        )
-    )
-    overdrive_client_key: str = FormField(
-        form=ConfigurationFormItem(
-            label=_("Client Key"),
-            type=ConfigurationFormItemType.TEXT,
-            description="The Overdrive client key.",
-            required=True,
-        )
-    )
-    overdrive_client_secret: str = FormField(
-        form=ConfigurationFormItem(
-            label=_("Client Secret"),
-            type=ConfigurationFormItemType.TEXT,
-            description="The Overdrive client secret.",
-            required=True,
-        )
-    )
-
-    overdrive_server_nickname: str = FormField(
-        default=OverdriveConstants.PRODUCTION_SERVERS,
-        form=ConfigurationFormItem(
-            label=_("Server family"),
-            type=ConfigurationFormItemType.SELECT,
-            required=False,
-            description="Unless you hear otherwise from Overdrive, your integration should use their production servers.",
-            options={
-                OverdriveConstants.PRODUCTION_SERVERS: ("Production"),
-                OverdriveConstants.TESTING_SERVERS: _("Testing"),
-            },
-        ),
-    )
-
-
-class OverdriveLibrarySettings(BaseCirculationEbookLoanSettings):
-    ils_name: str = FormField(
-        default=OverdriveConstants.ILS_NAME_DEFAULT,
-        form=ConfigurationFormItem(
-            label=_("ILS Name"),
-            description=_(
-                "When multiple libraries share an Overdrive account, Overdrive uses a setting called 'ILS Name' to determine which ILS to check when validating a given patron."
-            ),
-        ),
-    )
-
-
-class OverdriveChildSettings(BaseSettings):
-    external_account_id: str | None = FormField(
-        form=ConfigurationFormItem(
-            label=_("Library ID"),
-            required=True,
-        )
-    )
 
 
 class OverdriveToken(NamedTuple):
@@ -288,10 +137,6 @@ class OverdriveAPI(
         "PatronHasExceededCheckoutLimit": PatronLoanLimitReached,
         "PatronHasExceededCheckoutLimit_ForCPC": PatronLoanLimitReached,
     }
-
-    # An OverDrive defined constant indicating the "main" or parent account
-    # associated with an OverDrive collection.
-    OVERDRIVE_MAIN_ACCOUNT_ID = -1
 
     # A lock for threaded usage.
     lock = RLock()
@@ -386,7 +231,7 @@ class OverdriveAPI(
 
     @classmethod
     def label(cls) -> str:
-        return DataSourceConstants.OVERDRIVE
+        return OVERDRIVE_LABEL
 
     @classmethod
     def description(cls) -> str:
@@ -439,7 +284,7 @@ class OverdriveAPI(
         # This is set by an access to .collection_token
         self._collection_token: str | None = None
         self.overdrive_bibliographic_coverage_provider = (
-            OverdriveBibliographicCoverageProvider(collection, api_class=self)
+            OverdriveBibliographicCoverageProvider(collection, api=self)
         )
 
     @property
@@ -539,7 +384,7 @@ class OverdriveAPI(
             #
             # Instead of looking for the library ID itself in these
             # documents, we should look for the constant main account id.
-            return self.OVERDRIVE_MAIN_ACCOUNT_ID
+            return OVERDRIVE_MAIN_ACCOUNT_ID
         return int(self._library_id)
 
     def get(
@@ -701,7 +546,7 @@ class OverdriveAPI(
             collection_token=self.collection_token,
             sort="dateAdded:desc",
         )
-        return self.make_link_safe(url)
+        return _make_link_safe(url)
 
     def _get_book_list_page(
         self,
@@ -753,7 +598,7 @@ class OverdriveAPI(
             limit=str(self.PAGE_SIZE_LIMIT),
             collection_token=self.collection_token,
         )
-        next_link: str | None = self.make_link_safe(initial_next_link)
+        next_link: str | None = _make_link_safe(initial_next_link)
         while next_link:
             page_inventory, next_link = self._get_book_list_page(next_link)
             # We won't be sending out any events for these books yet,
@@ -771,31 +616,6 @@ class OverdriveAPI(
         )
         status_code, headers, content = self.get(url)
         return json.loads(content)  # type: ignore[no-any-return]
-
-    @classmethod
-    def make_link_safe(cls, url: str) -> str:
-        """Turn a server-provided link into a link the server will accept!
-
-        The {} part is completely obnoxious and I have complained about it to
-        Overdrive.
-
-        The availability part is to make sure we always use v2 of the
-        availability API, even if Overdrive sent us a link to v1.
-        """
-        parts = list(urlsplit(url))
-        parts[2] = quote(parts[2])
-        endings = ("/availability", "/availability/")
-        if parts[2].startswith("/v1/collections/") and any(
-            parts[2].endswith(x) for x in endings
-        ):
-            parts[2] = parts[2].replace("/v1/collections/", "/v2/collections/", 1)
-        query_string = parts[3]
-        query_string = query_string.replace("+", "%2B")
-        query_string = query_string.replace(":", "%3A")
-        query_string = query_string.replace("{", "%7B")
-        query_string = query_string.replace("}", "%7D")
-        parts[3] = query_string
-        return urlunsplit(tuple(parts))
 
     def _do_get(self, url: str, headers: dict[str, str], **kwargs: Any) -> Response:
         """This method is overridden in MockOverdriveAPI."""
@@ -1776,7 +1596,7 @@ class OverdriveAPI(
             circulation_link = book["availability_link"]
             # Make sure we use v2 of the availability API,
             # even if Overdrive gave us a link to v1.
-            circulation_link = self.make_link_safe(circulation_link)
+            circulation_link = _make_link_safe(circulation_link)
         return book, self.get(circulation_link, {})
 
     def update_formats(self, licensepool: LicensePool) -> None:
@@ -2002,1254 +1822,3 @@ class OverdriveAPI(
         else:
             link += "&contentfile=true"
         return link
-
-
-class OverdriveCirculationMonitor(CollectionMonitor, TimelineMonitor):
-    """Maintain LicensePools for recently changed Overdrive titles. Create
-    basic Editions for any new LicensePools that show up.
-    """
-
-    MAXIMUM_BOOK_RETRIES = 5
-    SERVICE_NAME = "Overdrive Circulation Monitor"
-    PROTOCOL = OverdriveAPI.label()
-    OVERLAP = datetime.timedelta(minutes=1)
-
-    @inject
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        api_class: type[OverdriveAPI] = OverdriveAPI,
-        analytics: Analytics = Provide[Services.analytics.analytics],
-    ) -> None:
-        """Constructor."""
-        super().__init__(_db, collection)
-        self.api = api_class(_db, collection)
-        self.analytics = analytics
-
-    def recently_changed_ids(
-        self, start: datetime.datetime, cutoff: datetime.datetime | None
-    ) -> Generator[dict[str, str]]:
-        return self.api.recently_changed_ids(start, cutoff)
-
-    def catch_up_from(
-        self,
-        start: datetime.datetime,
-        cutoff: datetime.datetime | None,
-        progress: TimestampData,
-    ) -> None:
-        """Find Overdrive books that changed recently.
-
-        :progress: A TimestampData representing the time previously
-            covered by this Monitor.
-        """
-        overdrive_data_source = DataSource.lookup(self._db, DataSource.OVERDRIVE)
-
-        # Ask for changes between the last time covered by the Monitor
-        # and the current time.
-        total_books = 0
-        for book in self.recently_changed_ids(start, cutoff):
-            total_books += 1
-            if not total_books % 100:
-                self.log.info("%s books processed", total_books)
-            if not book:
-                continue
-
-            book_changed = False
-            try:
-                book_changed = self.process_book(book, progress)
-                self._db.commit()
-            except Exception as e:
-                progress.exception = "".join(traceback.format_exception(e))
-
-            if self.should_stop(start, book, book_changed):
-                break
-
-        progress.achievements = "Books processed: %d." % total_books
-
-    @retry(
-        retry=(
-            retry_if_exception_type(StaleDataError)
-            | retry_if_exception_type(ObjectDeletedError)
-        ),
-        stop=stop_after_attempt(MAXIMUM_BOOK_RETRIES),
-        wait=wait_exponential(multiplier=1, min=1, max=60),
-        reraise=True,
-    )
-    def process_book(self, book: dict[str, Any], progress: TimestampData) -> bool:
-        # Attempt to create/update the book up to MAXIMUM_BOOK_RETRIES times.
-        try:
-            with self._db.begin_nested():
-                _, _, is_changed = self.api.update_licensepool(book)
-                book_changed = is_changed
-        except Exception:
-            self.log.exception("exception on update_licensepool: ")
-            raise
-        return book_changed
-
-    def should_stop(
-        self,
-        start: datetime.datetime,
-        api_description: dict[str, Any],
-        is_changed: bool,
-    ) -> bool | None:
-        pass
-
-
-class NewTitlesOverdriveCollectionMonitor(OverdriveCirculationMonitor):
-    """Monitor the Overdrive collection for newly added titles.
-    This catches any new titles that slipped through the cracks of the
-    RecentOverdriveCollectionMonitor.  This monitor queries for all titles
-    ordered by dateAdded in reverse chronological order.  However,
-    according to Overdrive (see https://ebce-lyrasis.atlassian.net/browse/PP-1002),
-    the dateAdded value is will not necessarily be in reverse chronological order.
-    If the collection this is being search is a consortial account and a title that
-    is present in a linked Advantage account has been added to the consortial collection
-    the dateAdded value will reflect the date it was added to the Advantage account. However,
-    it's position in the list will reflect the date it was added to the consortial account.
-    So ugly, but Overdrive insists this is the expected behavior.
-
-    To mitigate the possibility of missing new titles,  we can count the number of consecutive
-    titles with a dateAdded that is out of scope.  If we exceed that threshold, only then do
-    we stop processing.
-    """
-
-    SERVICE_NAME = "Overdrive New Title Monitor"
-    OVERLAP = datetime.timedelta(days=7)
-    DEFAULT_START_TIME = OverdriveCirculationMonitor.NEVER
-    MAX_CONSECUTIVE_OUT_OF_SCOPE_DATES = 1000
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self._consecutive_items_out_of_scope = 0
-
-    def recently_changed_ids(
-        self, start: datetime.datetime, cutoff: datetime.datetime | None
-    ) -> Generator[dict[str, str]]:
-        """Ignore the dates and return all IDs ordered by dateAdded in reverse chronological order."""
-        return self.api.all_ids()
-
-    def should_stop(
-        self,
-        start: datetime.datetime | None,
-        api_description: dict[str, Any],
-        is_changed: bool,
-    ) -> bool | None:
-        if not start or start is self.NEVER:
-            # This monitor has never run before. It should ask about
-            # every single book.
-            return False
-
-        # We should stop if this book was added before our start time.
-        date_added = api_description.get("date_added")
-        if not date_added:
-            # We don't know when this book was added -- shouldn't happen.
-            return False
-
-        try:
-            date_added = dateutil.parser.parse(date_added)
-        except ValueError as e:
-            # The date format is unparseable -- shouldn't happen.
-            self.log.error("Got invalid date: %s", date_added)
-            return False
-
-        date_out_of_scope = date_added < start
-        self.log.info(
-            f"Date added: {date_added}, start time: {start}, date out of scope: {date_out_of_scope}"
-        )
-
-        if date_out_of_scope:
-            self._consecutive_items_out_of_scope += 1
-            if (
-                self._consecutive_items_out_of_scope
-                > self.MAX_CONSECUTIVE_OUT_OF_SCOPE_DATES
-            ):
-                self.log.info(
-                    f"Max consecutive out of scope date threshold of {self.MAX_CONSECUTIVE_OUT_OF_SCOPE_DATES} "
-                    f"breached! We should stop now."
-                )
-                return True
-
-        else:
-            # reset consecutive counter
-            if self._consecutive_items_out_of_scope > 0:
-                self.log.info(
-                    f"We encountered a title that was added within our scope that followed a title that was out "
-                    f"of scope. Resetting counter from {self._consecutive_items_out_of_scope} consecutive items "
-                    f"back to zero."
-                )
-                self._consecutive_items_out_of_scope = 0
-
-        return False
-
-
-class OverdriveCollectionReaper(IdentifierSweepMonitor):
-    """Check for books that are in the local collection but have left our
-    Overdrive collection.
-    """
-
-    SERVICE_NAME = "Overdrive Collection Reaper"
-    PROTOCOL = OverdriveAPI.label()
-    DEFAULT_BATCH_SIZE = 10
-
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        api_class: type[OverdriveAPI] = OverdriveAPI,
-    ) -> None:
-        super().__init__(_db, collection)
-        self.api = api_class(_db, collection)
-
-    def process_item(self, identifier: Identifier) -> None:
-        self.api.update_licensepool(identifier.identifier)
-
-
-class RecentOverdriveCollectionMonitor(OverdriveCirculationMonitor):
-    """Monitor recently changed books in the Overdrive collection."""
-
-    SERVICE_NAME = "Reverse Chronological Overdrive Collection Monitor"
-
-    # Report successful completion upon finding this number of
-    # consecutive books in the Overdrive results whose LicensePools
-    # haven't changed since last time. Overdrive results are not in
-    # strict chronological order, but if you see 100 consecutive books
-    # that haven't changed, you're probably done.
-    MAXIMUM_CONSECUTIVE_UNCHANGED_BOOKS = 100
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self.consecutive_unchanged_books = 0
-
-    def should_stop(
-        self,
-        start: datetime.datetime,
-        api_description: dict[str, Any],
-        is_changed: bool,
-    ) -> bool | None:
-        if is_changed:
-            self.consecutive_unchanged_books = 0
-        else:
-            self.consecutive_unchanged_books += 1
-            if (
-                self.consecutive_unchanged_books
-                >= self.MAXIMUM_CONSECUTIVE_UNCHANGED_BOOKS
-            ):
-                # We're supposed to stop this run after finding a
-                # run of books that have not changed, and we have
-                # in fact seen that many consecutive unchanged
-                # books.
-                self.log.info(
-                    "Stopping at %d unchanged books.", self.consecutive_unchanged_books
-                )
-                return True
-        return False
-
-
-class OverdriveFormatSweep(IdentifierSweepMonitor):
-    """Check the current formats of every Overdrive book
-    in our collection.
-    """
-
-    SERVICE_NAME = "Overdrive Format Sweep"
-    DEFAULT_BATCH_SIZE = 10
-    PROTOCOL = OverdriveAPI.label()
-
-    def __init__(
-        self,
-        _db: Session,
-        collection: Collection,
-        api_class: type[OverdriveAPI] = OverdriveAPI,
-    ) -> None:
-        super().__init__(_db, collection)
-        self.api = api_class(_db, collection)
-
-    def process_item(self, identifier: Identifier) -> None:
-        pools = identifier.licensed_through
-        for pool in pools:
-            self.api.update_formats(pool)
-            # if there are multiple pools they should all have the same formats
-            # so we break after processing the first one
-            break
-
-
-class OverdriveRepresentationExtractor(LoggerMixin):
-    """Extract useful information from Overdrive's JSON representations."""
-
-    def __init__(self, api: OverdriveAPI) -> None:
-        """Constructor.
-
-        :param api: An OverdriveAPI object. This will be used when deciding
-        which portions of a JSON representation are relevant to the active
-        Overdrive collection.
-        """
-        self.library_id = api.advantage_library_id
-
-    @classmethod
-    def availability_link_list(cls, book_list: dict[str, Any]) -> list[dict[str, str]]:
-        """:return: A list of dictionaries with keys `id`, `title`, `availability_link`."""
-        l = []
-        if not "products" in book_list:
-            return []
-
-        products = book_list["products"]
-        for product in products:
-            if not "id" in product:
-                cls.logger().warning("No ID found in %r", product)
-                continue
-            book_id = product["id"]
-            data = dict(
-                id=book_id,
-                title=product.get("title"),
-                author_name=None,
-                date_added=product.get("dateAdded"),
-            )
-            if "primaryCreator" in product:
-                creator = product["primaryCreator"]
-                if creator.get("role") == "Author":
-                    data["author_name"] = creator.get("name")
-            links = product.get("links", [])
-            if "availability" in links:
-                link = links["availability"]["href"]
-                data["availability_link"] = OverdriveAPI.make_link_safe(link)
-            else:
-                logging.getLogger("Overdrive API").warning(
-                    "No availability link for %s", book_id
-                )
-            l.append(data)
-        return l
-
-    @classmethod
-    def link(cls, page: dict[str, Any], rel: str) -> str | None:
-        if "links" in page and rel in page["links"]:
-            raw_link = page["links"][rel]["href"]
-            link = OverdriveAPI.make_link_safe(raw_link)
-        else:
-            link = None
-        return link
-
-    format_data_for_overdrive_format: dict[
-        str, list[tuple[str, str | None]] | tuple[str, str | None]
-    ] = {
-        "ebook-pdf-adobe": (Representation.PDF_MEDIA_TYPE, DeliveryMechanism.ADOBE_DRM),
-        "ebook-pdf-open": (Representation.PDF_MEDIA_TYPE, DeliveryMechanism.NO_DRM),
-        "ebook-epub-adobe": (
-            Representation.EPUB_MEDIA_TYPE,
-            DeliveryMechanism.ADOBE_DRM,
-        ),
-        "ebook-epub-open": (Representation.EPUB_MEDIA_TYPE, DeliveryMechanism.NO_DRM),
-        "audiobook-mp3": ("application/x-od-media", DeliveryMechanism.OVERDRIVE_DRM),
-        "music-mp3": ("application/x-od-media", DeliveryMechanism.OVERDRIVE_DRM),
-        "ebook-overdrive": [
-            (
-                MediaTypes.OVERDRIVE_EBOOK_MANIFEST_MEDIA_TYPE,
-                DeliveryMechanism.LIBBY_DRM,
-            ),
-            (
-                DeliveryMechanism.STREAMING_TEXT_CONTENT_TYPE,
-                DeliveryMechanism.STREAMING_DRM,
-            ),
-        ],
-        "audiobook-overdrive": [
-            (
-                MediaTypes.OVERDRIVE_AUDIOBOOK_MANIFEST_MEDIA_TYPE,
-                DeliveryMechanism.LIBBY_DRM,
-            ),
-            (
-                DeliveryMechanism.STREAMING_AUDIO_CONTENT_TYPE,
-                DeliveryMechanism.STREAMING_DRM,
-            ),
-        ],
-        "video-streaming": (
-            DeliveryMechanism.STREAMING_VIDEO_CONTENT_TYPE,
-            DeliveryMechanism.STREAMING_DRM,
-        ),
-        "ebook-kindle": (
-            DeliveryMechanism.KINDLE_CONTENT_TYPE,
-            DeliveryMechanism.KINDLE_DRM,
-        ),
-        "periodicals-nook": (
-            DeliveryMechanism.NOOK_CONTENT_TYPE,
-            DeliveryMechanism.NOOK_DRM,
-        ),
-    }
-
-    # A mapping of the overdrive format name to end sample content type
-    # Overdrive samples are not DRM protected so the links should be
-    # stored as the end sample content type
-    sample_format_to_content_type = {
-        "ebook-overdrive": "text/html",
-        "audiobook-wma": "audio/x-ms-wma",
-        "audiobook-mp3": "audio/mpeg",
-        "audiobook-overdrive": "text/html",
-        "ebook-epub-adobe": "application/epub+zip",
-        "magazine-overdrive": "text/html",
-    }
-
-    @classmethod
-    def internal_formats(
-        cls, overdrive_format: str
-    ) -> Generator[tuple[str, str | None]]:
-        """Yield all internal formats for the given Overdrive format.
-
-        Some Overdrive formats become multiple internal formats.
-
-        :yield: A sequence of (content type, DRM system) 2-tuples
-        """
-        result = cls.format_data_for_overdrive_format.get(overdrive_format)
-        if not result:
-            return
-        if isinstance(result, list):
-            yield from result
-        else:
-            yield result
-
-    ignorable_overdrive_formats: set[str] = set()
-
-    overdrive_role_to_simplified_role = {
-        "actor": Contributor.Role.ACTOR,
-        "artist": Contributor.Role.ARTIST,
-        "book producer": Contributor.Role.PRODUCER,
-        "associated name": Contributor.Role.ASSOCIATED,
-        "author": Contributor.Role.AUTHOR,
-        "author of introduction": Contributor.Role.INTRODUCTION,
-        "author of foreword": Contributor.Role.FOREWORD,
-        "author of afterword": Contributor.Role.AFTERWORD,
-        "contributor": Contributor.Role.CONTRIBUTOR,
-        "colophon": Contributor.Role.COLOPHON,
-        "adapter": Contributor.Role.ADAPTER,
-        "etc.": Contributor.Role.UNKNOWN,
-        "cast member": Contributor.Role.ACTOR,
-        "collaborator": Contributor.Role.COLLABORATOR,
-        "compiler": Contributor.Role.COMPILER,
-        "composer": Contributor.Role.COMPOSER,
-        "copyright holder": Contributor.Role.COPYRIGHT_HOLDER,
-        "director": Contributor.Role.DIRECTOR,
-        "editor": Contributor.Role.EDITOR,
-        "engineer": Contributor.Role.ENGINEER,
-        "executive producer": Contributor.Role.EXECUTIVE_PRODUCER,
-        "illustrator": Contributor.Role.ILLUSTRATOR,
-        "musician": Contributor.Role.MUSICIAN,
-        "narrator": Contributor.Role.NARRATOR,
-        "other": Contributor.Role.UNKNOWN,
-        "performer": Contributor.Role.PERFORMER,
-        "producer": Contributor.Role.PRODUCER,
-        "translator": Contributor.Role.TRANSLATOR,
-        "photographer": Contributor.Role.PHOTOGRAPHER,
-        "lyricist": Contributor.Role.LYRICIST,
-        "transcriber": Contributor.Role.TRANSCRIBER,
-        "designer": Contributor.Role.DESIGNER,
-    }
-
-    overdrive_medium_to_simplified_medium = {
-        "eBook": Edition.BOOK_MEDIUM,
-        "Video": Edition.VIDEO_MEDIUM,
-        "Audiobook": Edition.AUDIO_MEDIUM,
-        "Music": Edition.MUSIC_MEDIUM,
-        "Periodicals": Edition.PERIODICAL_MEDIUM,
-    }
-
-    DATE_FORMAT = "%Y-%m-%d"
-
-    @classmethod
-    def parse_roles(cls, id: str, rolestring: str) -> list[Contributor.Role]:
-        rolestring = rolestring.lower()
-        roles = [x.strip() for x in rolestring.split(",")]
-        if " and " in roles[-1]:
-            roles = roles[:-1] + [x.strip() for x in roles[-1].split(" and ")]
-        processed = []
-        for x in roles:
-            if x not in cls.overdrive_role_to_simplified_role:
-                cls.logger().error("Could not process role %s for %s", x, id)
-            else:
-                processed.append(cls.overdrive_role_to_simplified_role[x])
-        return processed
-
-    def book_info_to_circulation(self, book: dict[str, Any]) -> CirculationData:
-        """Note:  The json data passed into this method is from a different file/stream
-        from the json data that goes into the book_info_to_metadata() method.
-        """
-        # In Overdrive, 'reserved' books show up as books on
-        # hold. There is no separate notion of reserved books.
-        licenses_reserved = 0
-
-        licenses_owned = None
-        licenses_available = None
-        patrons_in_hold_queue = None
-
-        # TODO: The only reason this works for a NotFound error is the
-        # circulation code sticks the known book ID into `book` ahead
-        # of time. That's a code smell indicating that this system
-        # needs to be refactored.
-        if "reserveId" in book and not "id" in book:
-            book["id"] = book["reserveId"]
-        if not "id" in book:
-            self.log.error("Book has no ID: %r", book)
-            raise PalaceValueError("Book must have an id to be processed")
-        overdrive_id = book["id"]
-        primary_identifier = IdentifierData(Identifier.OVERDRIVE_ID, overdrive_id)
-        # TODO: We might be able to use this information to avoid the
-        # need for explicit configuration of Advantage collections, or
-        # at least to keep Advantage collections more up-to-date than
-        # they would be otherwise, as a side effect of updating
-        # regular Overdrive collections.
-
-        # TODO: this would be the place to handle simultaneous use
-        # titles -- these can be detected with
-        # availabilityType="AlwaysAvailable" and have their
-        # .licenses_owned set to LicensePool.UNLIMITED_ACCESS.
-        # see http://developer.overdrive.com/apis/library-availability-new
-
-        # TODO: Cost-per-circ titles
-        # (availabilityType="LimitedAvailablility") can be handled
-        # similarly, though those can abruptly become unavailable, so
-        # UNLIMITED_ACCESS is probably not appropriate.
-
-        error_code = book.get("errorCode")
-        # TODO: It's not clear what other error codes there might be.
-        # The current behavior will respond to errors other than
-        # NotFound by leaving the book alone, but this might not be
-        # the right behavior.
-        if error_code == "NotFound":
-            licenses_owned = 0
-            licenses_available = 0
-            patrons_in_hold_queue = 0
-        elif book.get("isOwnedByCollections") is not False:
-            # We own this book.
-            licenses_owned = 0
-            licenses_available = 0
-
-            for account in self._get_applicable_accounts(book.get("accounts", [])):
-                licenses_owned += int(account.get("copiesOwned", 0))
-                licenses_available += int(account.get("copiesAvailable", 0))
-
-            if "numberOfHolds" in book:
-                if patrons_in_hold_queue is None:
-                    patrons_in_hold_queue = 0
-                patrons_in_hold_queue += book["numberOfHolds"]
-
-        return CirculationData(
-            data_source=DataSource.OVERDRIVE,
-            primary_identifier=primary_identifier,
-            licenses_owned=licenses_owned,
-            licenses_available=licenses_available,
-            licenses_reserved=licenses_reserved,
-            patrons_in_hold_queue=patrons_in_hold_queue,
-        )
-
-    def _get_applicable_accounts(
-        self, accounts: list[dict[str, Any]]
-    ) -> list[dict[str, Any]]:
-        """
-        Returns those accounts from the accounts array that apply the
-        current overdrive collection context.
-
-        If this is an overdrive parent collection, we want to return accounts
-        associated with the main OverDrive "library" and any non-main account
-        with sharing enabled.
-
-        If this is a child OverDrive collection, then we return only the
-        account associated with that child's OverDrive Advantage "library".
-        Additionally, we want to exclude the account if it is "shared" since
-        we will be counting it with the parent collection.
-        """
-
-        if self.library_id == OverdriveAPI.OVERDRIVE_MAIN_ACCOUNT_ID:
-            # this is a parent collection
-            filtered_result = filter(
-                lambda account: account.get("id")
-                == OverdriveAPI.OVERDRIVE_MAIN_ACCOUNT_ID
-                or account.get("shared", False),
-                accounts,
-            )
-        else:
-            # this is child collection
-            filtered_result = filter(
-                lambda account: account.get("id") == self.library_id
-                and not account.get("shared", False),
-                accounts,
-            )
-
-        return list(filtered_result)
-
-    @classmethod
-    def image_link_to_linkdata(cls, link: dict[str, str], rel: str) -> LinkData | None:
-        if not link or not "href" in link:
-            return None
-        href = link["href"]
-        if "00000000-0000-0000-0000" in href:
-            # This is a stand-in cover for preorders. It's better not
-            # to have a cover at all -- we might be able to get one
-            # later, or from another source.
-            return None
-        href = OverdriveAPI.make_link_safe(href)
-        media_type = link.get("type", None)
-        return LinkData(rel=rel, href=href, media_type=media_type)
-
-    @classmethod
-    def book_info_to_metadata(
-        cls,
-        book: dict[str, Any],
-        include_bibliographic: bool = True,
-        include_formats: bool = True,
-    ) -> Metadata | None:
-        """Turn Overdrive's JSON representation of a book into a Metadata
-        object.
-
-        Note:  The json data passed into this method is from a different file/stream
-        from the json data that goes into the book_info_to_circulation() method.
-        """
-        if not "id" in book:
-            return None
-        overdrive_id = book["id"]
-        primary_identifier = IdentifierData(Identifier.OVERDRIVE_ID, overdrive_id)
-
-        # If we trust classification data, we'll give it this weight.
-        # Otherwise we'll probably give it a fraction of this weight.
-        trusted_weight = Classification.TRUSTED_DISTRIBUTOR_WEIGHT
-
-        duration: int | None = None
-
-        if include_bibliographic:
-            title = book.get("title", None)
-            sort_title = book.get("sortTitle")
-            subtitle = book.get("subtitle", None)
-            series = book.get("series", None)
-            publisher = book.get("publisher", None)
-            imprint = book.get("imprint", None)
-
-            if "publishDate" in book:
-                published = strptime_utc(book["publishDate"][:10], cls.DATE_FORMAT)
-            else:
-                published = None
-
-            languages = [l["code"] for l in book.get("languages", [])]
-            if "eng" in languages or not languages:
-                language = "eng"
-            else:
-                language = sorted(languages)[0]
-
-            contributors = []
-            for creator in book.get("creators", []):
-                sort_name = creator["fileAs"]
-                display_name = creator["name"]
-                role = creator["role"]
-                roles = cls.parse_roles(overdrive_id, role) or [
-                    Contributor.Role.UNKNOWN
-                ]
-                contributor = ContributorData(
-                    sort_name=sort_name,
-                    display_name=display_name,
-                    roles=roles,
-                    biography=creator.get("bioText", None),
-                )
-                contributors.append(contributor)
-
-            subjects = []
-            for sub in book.get("subjects", []):
-                subject = SubjectData(
-                    type=Subject.OVERDRIVE,
-                    identifier=sub["value"],
-                    weight=trusted_weight,
-                )
-                subjects.append(subject)
-
-            for sub in book.get("keywords", []):
-                subject = SubjectData(
-                    type=Subject.TAG,
-                    identifier=sub["value"],
-                    # We don't use TRUSTED_DISTRIBUTOR_WEIGHT because
-                    # we don't know where the tags come from --
-                    # probably Overdrive users -- and they're
-                    # frequently wrong.
-                    weight=1,
-                )
-                subjects.append(subject)
-
-            extra = dict()
-            if "grade_levels" in book:
-                # n.b. Grade levels are measurements of reading level, not
-                # age appropriateness. We can use them as a measure of age
-                # appropriateness in a pinch, but we weight them less
-                # heavily than TRUSTED_DISTRIBUTOR_WEIGHT.
-                for i in book["grade_levels"]:
-                    subject = SubjectData(
-                        type=Subject.GRADE_LEVEL,
-                        identifier=i["value"],
-                        weight=round(trusted_weight / 10),
-                    )
-                    subjects.append(subject)
-
-            overdrive_medium = book.get("mediaType", None)
-            if (
-                overdrive_medium
-                and overdrive_medium not in cls.overdrive_medium_to_simplified_medium
-            ):
-                cls.logger().error(
-                    "Could not process medium %s for %s", overdrive_medium, overdrive_id
-                )
-
-            medium = cls.overdrive_medium_to_simplified_medium.get(
-                overdrive_medium, Edition.BOOK_MEDIUM
-            )
-
-            measurements = []
-            if "awards" in book:
-                extra["awards"] = book.get("awards", [])
-                num_awards = len(extra["awards"])
-                measurements.append(
-                    MeasurementData(Measurement.AWARDS, str(num_awards))
-                )
-
-            for name, subject_type in (
-                ("ATOS", Subject.ATOS_SCORE),
-                ("lexileScore", Subject.LEXILE_SCORE),
-                ("interestLevel", Subject.INTEREST_LEVEL),
-            ):
-                if not name in book:
-                    continue
-                identifier = str(book[name])
-                subjects.append(
-                    SubjectData(
-                        type=subject_type, identifier=identifier, weight=trusted_weight
-                    )
-                )
-
-            for grade_level_info in book.get("gradeLevels", []):
-                grade_level = grade_level_info.get("value")
-                subjects.append(
-                    SubjectData(
-                        type=Subject.GRADE_LEVEL,
-                        identifier=grade_level,
-                        weight=trusted_weight,
-                    )
-                )
-
-            identifiers = []
-            links = []
-            sample_hrefs = set()
-            for format in book.get("formats", []):
-                duration_str: str | None = format.get("duration")
-                if duration_str is not None:
-                    # Using this method only the last valid duration attribute is captured
-                    # If there are multiple formats with different durations, the edition will ignore the rest
-                    try:
-                        hrs, mins, secs = duration_str.split(":")
-                        duration = (int(hrs) * 3600) + (int(mins) * 60) + int(secs)
-                    except Exception as ex:
-                        cls.logger().error(
-                            f"Duration ({duration_str}) could not be parsed: {ex}"
-                        )
-
-                for new_id in format.get("identifiers", []):
-                    t = new_id["type"]
-                    v = new_id["value"]
-                    orig_v = v
-                    type_key = None
-                    if t == "ASIN":
-                        type_key = Identifier.ASIN
-                    elif t == "ISBN":
-                        type_key = Identifier.ISBN
-                        if len(v) == 10:
-                            v = isbnlib.to_isbn13(v)
-                        if v is None or not isbnlib.is_isbn13(v):
-                            # Overdrive sometimes uses invalid values
-                            # like "n/a" as placeholders. Ignore such
-                            # values to avoid a situation where hundreds of
-                            # books appear to have the same ISBN. ISBNs
-                            # which fail check digit checks or are invalid
-                            # also can occur. Log them for review.
-                            cls.logger().info("Bad ISBN value provided: %s", orig_v)
-                            continue
-                    elif t == "DOI":
-                        type_key = Identifier.DOI
-                    elif t == "UPC":
-                        type_key = Identifier.UPC
-                    elif t == "PublisherCatalogNumber":
-                        continue
-                    if type_key and v:
-                        identifiers.append(IdentifierData(type_key, v, 1))
-
-                # Samples become links.
-                if "samples" in format:
-                    for sample_info in format["samples"]:
-                        href = sample_info["url"]
-                        # Have we already parsed this sample? Overdrive repeats samples per format
-                        if href in sample_hrefs:
-                            continue
-
-                        # Every sample has its own format type
-                        overdrive_format_name = sample_info.get("formatType")
-                        if not overdrive_format_name:
-                            # Malformed sample
-                            continue
-                        content_type = cls.sample_format_to_content_type.get(
-                            overdrive_format_name
-                        )
-                        if not content_type:
-                            # Unusable by us.
-                            cls.logger().warning(
-                                f"Did not find a sample format mapping for '{overdrive_format_name}': {href}"
-                            )
-                            continue
-
-                        if Representation.is_media_type(content_type):
-                            links.append(
-                                LinkData(
-                                    rel=Hyperlink.SAMPLE,
-                                    href=href,
-                                    media_type=content_type,
-                                )
-                            )
-                            sample_hrefs.add(href)
-
-            # A cover and its thumbnail become a single LinkData.
-            if "images" in book:
-                images = book["images"]
-                image_data = cls.image_link_to_linkdata(
-                    images.get("cover"), Hyperlink.IMAGE
-                )
-                for name in ["cover300Wide", "cover150Wide", "thumbnail"]:
-                    # Try to get a thumbnail that's as close as possible
-                    # to the size we use.
-                    image = images.get(name)
-                    thumbnail_data = cls.image_link_to_linkdata(
-                        image, Hyperlink.THUMBNAIL_IMAGE
-                    )
-                    if not image_data:
-                        image_data = cls.image_link_to_linkdata(image, Hyperlink.IMAGE)
-                    if thumbnail_data:
-                        break
-
-                if image_data:
-                    if thumbnail_data:
-                        image_data.thumbnail = thumbnail_data
-                    links.append(image_data)
-
-            # Descriptions become links.
-            short = book.get("shortDescription")
-            full = book.get("fullDescription")
-            if full:
-                links.append(
-                    LinkData(
-                        rel=Hyperlink.DESCRIPTION,
-                        content=full,
-                        media_type="text/html",
-                    )
-                )
-
-            if short and (not full or not full.startswith(short)):
-                links.append(
-                    LinkData(
-                        rel=Hyperlink.SHORT_DESCRIPTION,
-                        content=short,
-                        media_type="text/html",
-                    )
-                )
-
-            # Add measurements: rating and popularity
-            if book.get("starRating") is not None and book["starRating"] > 0:
-                measurements.append(
-                    MeasurementData(
-                        quantity_measured=Measurement.RATING, value=book["starRating"]
-                    )
-                )
-
-            if book.get("popularity"):
-                measurements.append(
-                    MeasurementData(
-                        quantity_measured=Measurement.POPULARITY,
-                        value=book["popularity"],
-                    )
-                )
-
-            metadata = Metadata(
-                data_source=DataSource.OVERDRIVE,
-                title=title,
-                subtitle=subtitle,
-                sort_title=sort_title,
-                language=language,
-                medium=medium,
-                series=series,
-                publisher=publisher,
-                imprint=imprint,
-                published=published,
-                primary_identifier=primary_identifier,
-                identifiers=identifiers,
-                subjects=subjects,
-                contributors=contributors,
-                measurements=measurements,
-                links=links,
-                duration=duration,
-            )
-        else:
-            metadata = Metadata(
-                data_source=DataSource.OVERDRIVE,
-                primary_identifier=primary_identifier,
-            )
-
-        if include_formats:
-            formats = []
-            for format in book.get("formats", []):
-                format_id = format["id"]
-                internal_formats = list(cls.internal_formats(format_id))
-                if internal_formats:
-                    for content_type, drm_scheme in internal_formats:
-                        formats.append(FormatData(content_type, drm_scheme))
-                elif format_id not in cls.ignorable_overdrive_formats:
-                    cls.logger().error(
-                        "Could not process Overdrive format %s for %s",
-                        format_id,
-                        overdrive_id,
-                    )
-
-            # Also make a CirculationData so we can write the formats,
-            circulationdata = CirculationData(
-                data_source=DataSource.OVERDRIVE,
-                primary_identifier=primary_identifier,
-                formats=formats,
-            )
-
-            metadata.circulation = circulationdata
-
-        return metadata
-
-
-class OverdriveAdvantageAccount:
-    """Holder and parser for data associated with Overdrive Advantage."""
-
-    def __init__(
-        self, parent_library_id: str, library_id: str, name: str, token: str
-    ) -> None:
-        """Constructor.
-
-        :param parent_library_id: The library ID of the parent Overdrive
-            account.
-        :param library_id: The library ID of the Overdrive Advantage account.
-        :param name: The name of the library whose Advantage account this is.
-        :param token: The collection token for this Advantage account
-        """
-        self.parent_library_id = parent_library_id
-        self.library_id = library_id
-        self.name = name
-        self.token = token
-
-    @classmethod
-    def from_representation(cls, content: str) -> Generator[Self]:
-        """Turn the representation of an advantageAccounts link into a list of
-        OverdriveAdvantageAccount objects.
-
-        :param content: The data obtained by following an advantageAccounts
-            link.
-        :yield: A sequence of OverdriveAdvantageAccount objects.
-        """
-        data = json.loads(content)
-        parent_id = str(data.get("id"))
-        accounts = data.get("advantageAccounts", {})
-        for account in accounts:
-            name = account["name"]
-            products_link = account["links"]["products"]["href"]
-            library_id = str(account.get("id"))
-            name = account.get("name")
-            token = account.get("collectionToken")
-            yield cls(
-                parent_library_id=parent_id,
-                library_id=library_id,
-                name=name,
-                token=token,
-            )
-
-    def to_collection(self, _db: Session) -> tuple[Collection, Collection]:
-        """Find or create a Collection object for this Overdrive Advantage
-        account.
-
-        :return: a 2-tuple of Collections (primary Overdrive
-            collection, Overdrive Advantage collection)
-        """
-        # First find the parent Collection.
-        parent = _db.execute(
-            select(Collection)
-            .join(IntegrationConfiguration)
-            .where(
-                IntegrationConfiguration.protocol == OverdriveAPI.label(),
-                IntegrationConfiguration.goal == Goals.LICENSE_GOAL,
-                IntegrationConfiguration.settings_dict.contains(
-                    {"external_account_id": self.parent_library_id}
-                ),
-            )
-        ).scalar_one_or_none()
-        if parent is None:
-            # Without the parent's credentials we can't access the child.
-            raise ValueError(
-                "Cannot create a Collection whose parent does not already exist."
-            )
-        name = parent.name + " / " + self.name
-        child = _db.execute(
-            select(Collection)
-            .join(IntegrationConfiguration)
-            .where(
-                Collection.parent_id == parent.id,
-                IntegrationConfiguration.protocol == OverdriveAPI.label(),
-                IntegrationConfiguration.goal == Goals.LICENSE_GOAL,
-                IntegrationConfiguration.settings_dict.contains(
-                    {"external_account_id" == self.library_id}
-                ),
-            )
-        ).scalar_one_or_none()
-
-        if child is None:
-            # The child doesn't exist yet. Create it.
-            child, _ = Collection.by_name_and_protocol(_db, name, OverdriveAPI.label())
-            child.parent = parent
-            child_settings = OverdriveChildSettings.model_construct(
-                external_account_id=self.library_id
-            )
-            integration_settings_update(
-                OverdriveChildSettings, child.integration_configuration, child_settings
-            )
-        else:
-            # Set or update the name of the collection to reflect the name of
-            # the library, just in case that name has changed.
-            child.integration_configuration.name = name
-
-        return parent, child
-
-
-class OverdriveBibliographicCoverageProvider(BibliographicCoverageProvider):
-    """Fill in bibliographic metadata for Overdrive records.
-
-    This will occasionally fill in some availability information for a
-    single Collection, but we rely on Monitors to keep availability
-    information up to date for all Collections.
-    """
-
-    SERVICE_NAME = "Overdrive Bibliographic Coverage Provider"
-    DATA_SOURCE_NAME = DataSource.OVERDRIVE
-    PROTOCOL = OverdriveAPI.label()
-    INPUT_IDENTIFIER_TYPES = Identifier.OVERDRIVE_ID
-
-    def __init__(
-        self,
-        collection: Collection,
-        api_class: type[OverdriveAPI] | OverdriveAPI = OverdriveAPI,
-        **kwargs: Any,
-    ) -> None:
-        """Constructor.
-
-        :param collection: Provide bibliographic coverage to all
-            Overdrive books in the given Collection.
-        :param api_class: Instantiate this class with the given Collection,
-            rather than instantiating OverdriveAPI.
-        """
-        super().__init__(collection, **kwargs)
-        if isinstance(api_class, OverdriveAPI):
-            # Use a previously instantiated OverdriveAPI instance
-            # rather than creating a new one.
-            self.api = api_class
-        else:
-            # A web application should not use this option because it
-            # will put a non-scoped session in the mix.
-            _db = Session.object_session(collection)
-            self.api = api_class(_db, collection)
-
-    def process_item(self, identifier: Identifier) -> Identifier | CoverageFailure:
-        info = self.api.metadata_lookup(identifier)
-        error = None
-        if info.get("errorCode") == "NotFound":
-            error = "ID not recognized by Overdrive: %s" % identifier.identifier
-        elif info.get("errorCode") == "InvalidGuid":
-            error = "Invalid Overdrive ID: %s" % identifier.identifier
-
-        if error:
-            return self.failure(identifier, error, transient=False)  # type: ignore[no-any-return]
-
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-
-        if not metadata:
-            e = "Could not extract metadata from Overdrive data: %r" % info
-            return self.failure(identifier, e)  # type: ignore[no-any-return]
-
-        self.metadata_pre_hook(metadata)
-        return self.set_metadata(identifier, metadata)  # type: ignore[no-any-return]
-
-    def metadata_pre_hook(self, metadata: Metadata) -> Metadata:
-        """A hook method that allows subclasses to modify a Metadata
-        object derived from Overdrive before it's applied.
-        """
-        return metadata
-
-
-class GenerateOverdriveAdvantageAccountList(InputScript):
-    """Generates a CSV containing the following fields:
-    circulation manager
-    collection
-    client_key
-    external_account_id
-    library_token
-    advantage_name
-    advantage_id
-    advantage_token
-    already_configured
-    """
-
-    def __init__(self, _db: Session | None = None, *args: Any, **kwargs: Any) -> None:
-        super().__init__(_db, *args, **kwargs)
-        self._data: list[list[str | bool]] = list()
-
-    def _create_overdrive_api(self, collection: Collection) -> OverdriveAPI:
-        return OverdriveAPI(_db=self._db, collection=collection)
-
-    def do_run(self, **kwargs: Any) -> None:
-        parsed = GenerateOverdriveAdvantageAccountList.parse_command_line(
-            _db=self._db, **kwargs
-        )
-        query = Collection.by_protocol(self._db, protocol=OverdriveAPI.label())
-        for collection in query.filter(Collection.parent_id == None):
-            api = self._create_overdrive_api(collection=collection)
-            client_key = api.client_key().decode()
-            client_secret = api.client_secret().decode()
-            library_id = api.library_id()
-
-            try:
-                library_token = api.collection_token
-                advantage_accounts = api.get_advantage_accounts()
-
-                for aa in advantage_accounts:
-                    existing_child_collections = query.filter(
-                        Collection.parent_id == collection.id
-                    )
-                    already_configured_aa_libraries = [
-                        OverdriveAPI.child_settings_load(
-                            e.integration_configuration
-                        ).external_account_id
-                        for e in existing_child_collections
-                    ]
-                    self._data.append(
-                        [
-                            collection.name,
-                            library_id,
-                            client_key,
-                            client_secret,
-                            library_token,
-                            aa.name,
-                            aa.library_id,
-                            aa.token,
-                            aa.library_id in already_configured_aa_libraries,
-                        ]
-                    )
-            except Exception as e:
-                logging.error(
-                    f"Could not connect to collection {collection.name}: reason: {str(e)}."
-                )
-
-        file_path = parsed.output_file_path[0]
-        circ_manager_name = parsed.circulation_manager_name[0]
-        self.write_csv(output_file_path=file_path, circ_manager_name=circ_manager_name)
-
-    def write_csv(self, output_file_path: str, circ_manager_name: str) -> None:
-        with open(output_file_path, "w", newline="") as csvfile:
-            writer = csv.writer(csvfile)
-            writer.writerow(
-                [
-                    "cm",
-                    "collection",
-                    "overdrive_library_id",
-                    "client_key",
-                    "client_secret",
-                    "library_token",
-                    "advantage_name",
-                    "advantage_id",
-                    "advantage_token",
-                    "already_configured",
-                ]
-            )
-            for i in self._data:
-                i.insert(0, circ_manager_name)
-                writer.writerow(i)
-
-    @classmethod
-    def arg_parser(cls) -> argparse.ArgumentParser:
-        parser = argparse.ArgumentParser()
-        parser.add_argument(
-            "--output-file-path",
-            help="The path of an output file",
-            metavar="o",
-            nargs=1,
-        )
-
-        parser.add_argument(
-            "--circulation-manager-name",
-            help="The name of the circulation-manager",
-            metavar="c",
-            nargs=1,
-            required=True,
-        )
-
-        parser.add_argument(
-            "--file-format",
-            help="The file format of the output file",
-            metavar="f",
-            nargs=1,
-            default="csv",
-        )
-
-        return parser
-
-
-class OverdriveAdvantageAccountListScript(Script):
-    def run(self) -> None:
-        """Explain every Overdrive collection and, for each one, all of its
-        Advantage collections.
-        """
-        collections = Collection.by_protocol(self._db, OverdriveAPI.label())
-        for collection in collections:
-            self.explain_main_collection(collection)
-            print()
-
-    def explain_main_collection(self, collection: Collection) -> None:
-        """Explain an Overdrive collection and all of its Advantage
-        collections.
-        """
-        api = OverdriveAPI(self._db, collection)
-        print("Main Overdrive collection: %s" % collection.name)
-        print("\n".join(collection.explain()))
-        print("A few of the titles in the main collection:")
-        for i, book in enumerate(api.all_ids()):
-            print("", book["title"])
-            if i > 10:
-                break
-        advantage_accounts = list(api.get_advantage_accounts())
-        print("%d associated Overdrive Advantage account(s)." % len(advantage_accounts))
-        for advantage_collection in advantage_accounts:
-            self.explain_advantage_collection(advantage_collection)
-            print()
-
-    def explain_advantage_collection(
-        self, collection: OverdriveAdvantageAccount
-    ) -> None:
-        """Explain a single Overdrive Advantage collection."""
-        parent_collection, child = collection.to_collection(self._db)
-        print(" Overdrive Advantage collection: %s" % child.name)
-        print(" " + ("\n ".join(child.explain())))
-        print(" A few of the titles in this Advantage collection:")
-        child_api = OverdriveAPI(self._db, child)
-        for i, book in enumerate(child_api.all_ids()):
-            print(" ", book["title"])
-            if i > 10:
-                break
-
-
-class OverdriveManifestFulfillment(RedirectFulfillment):
-    def __init__(self, content_link: str, scope_string: str, access_token: str) -> None:
-        super().__init__(content_link)
-        self.scope_string = scope_string
-        self.access_token = access_token
-
-    def response(self) -> flask.Response:
-        headers = {
-            "Location": self.content_link,
-            "X-Overdrive-Scope": self.scope_string,
-            "X-Overdrive-Patron-Authorization": f"Bearer {self.access_token}",
-            "Content-Type": "text/plain",
-        }
-        return flask.Response("", 302, headers)
