+++ tests/manager/api/overdrive/test_api.py
@@ -1,22 +1,17 @@
 from __future__ import annotations
 
 import base64
-import csv
 import json
-import logging
 import os
 import random
-from dataclasses import dataclass
 from datetime import timedelta
-from typing import TYPE_CHECKING, Any, cast
+from typing import TYPE_CHECKING, Any
 from unittest.mock import MagicMock, create_autospec, patch
 
 import pytest
 from requests import Response
-from sqlalchemy.orm.exc import ObjectDeletedError, StaleDataError
 
 from palace.manager.api.circulation import (
-    CirculationAPI,
     FetchFulfillment,
     Fulfillment,
     HoldInfo,
@@ -38,31 +33,12 @@ from palace.manager.api.circulation_exceptions import (
     PatronLoanLimitReached,
 )
 from palace.manager.api.config import Configuration
-from palace.manager.api.overdrive import (
-    GenerateOverdriveAdvantageAccountList,
-    NewTitlesOverdriveCollectionMonitor,
-    OverdriveAdvantageAccount,
-    OverdriveAPI,
-    OverdriveBibliographicCoverageProvider,
-    OverdriveCirculationMonitor,
-    OverdriveCollectionReaper,
-    OverdriveConstants,
-    OverdriveFormatSweep,
-    OverdriveManifestFulfillment,
-    OverdriveRepresentationExtractor,
-    RecentOverdriveCollectionMonitor,
-)
+from palace.manager.api.overdrive.api import OverdriveAPI
+from palace.manager.api.overdrive.constants import OverdriveConstants
+from palace.manager.api.overdrive.fulfillment import OverdriveManifestFulfillment
 from palace.manager.core.config import CannotLoadConfiguration
-from palace.manager.core.coverage import CoverageFailure
-from palace.manager.core.exceptions import BasePalaceException, PalaceValueError
-from palace.manager.core.metadata_layer import LinkData, TimestampData
-from palace.manager.integration.goals import Goals
-from palace.manager.scripts.coverage_provider import RunCollectionCoverageProviderScript
-from palace.manager.service.analytics.analytics import Analytics
+from palace.manager.core.exceptions import BasePalaceException
 from palace.manager.sqlalchemy.constants import MediaTypes
-from palace.manager.sqlalchemy.model.classification import Subject
-from palace.manager.sqlalchemy.model.collection import Collection
-from palace.manager.sqlalchemy.model.contributor import Contributor
 from palace.manager.sqlalchemy.model.datasource import DataSource
 from palace.manager.sqlalchemy.model.edition import Edition
 from palace.manager.sqlalchemy.model.identifier import Identifier
@@ -72,80 +48,18 @@ from palace.manager.sqlalchemy.model.licensing import (
     LicensePoolDeliveryMechanism,
     RightsStatus,
 )
-from palace.manager.sqlalchemy.model.measurement import Measurement
-from palace.manager.sqlalchemy.model.patron import Patron
-from palace.manager.sqlalchemy.model.resource import Hyperlink, Representation
+from palace.manager.sqlalchemy.model.resource import Representation
 from palace.manager.util.datetime_helpers import datetime_utc, utc_now
 from palace.manager.util.http import BadResponseException
 from tests.fixtures.database import DatabaseTransactionFixture
-from tests.fixtures.files import FilesFixture
 from tests.fixtures.library import LibraryFixture
+from tests.fixtures.overdrive import OverdriveAPIFixture
 from tests.fixtures.webserver import MockAPIServer, MockAPIServerResponse
 from tests.mocks.mock import MockHTTPClient, MockRequestsResponse
 from tests.mocks.overdrive import MockOverdriveAPI
 
 if TYPE_CHECKING:
-    from tests.fixtures.time import Time
-
-
-class OverdriveFilesFixture(FilesFixture):
-    """A fixture providing access to Overdrive files."""
-
-    def __init__(self):
-        super().__init__("overdrive")
-
-
-@pytest.fixture()
-def overdrive_files_fixture() -> OverdriveFilesFixture:
-    """A fixture providing access to Overdrive files."""
-    return OverdriveFilesFixture()
-
-
-class OverdriveAPIFixture:
-    def __init__(self, db: DatabaseTransactionFixture, data: OverdriveFilesFixture):
-        self.db = db
-        self.data = data
-        library = db.default_library()
-        self.collection = MockOverdriveAPI.mock_collection(
-            db.session, db.default_library()
-        )
-        self.api = MockOverdriveAPI(db.session, self.collection)
-        self.circulation = CirculationAPI(
-            db.session,
-            library,
-            {self.collection.id: self.api},
-        )
-        os.environ[
-            f"{Configuration.OD_PREFIX_TESTING_PREFIX}_{Configuration.OD_FULFILLMENT_CLIENT_KEY_SUFFIX}"
-        ] = "TestingKey"
-        os.environ[
-            f"{Configuration.OD_PREFIX_TESTING_PREFIX}_{Configuration.OD_FULFILLMENT_CLIENT_SECRET_SUFFIX}"
-        ] = "TestingSecret"
-
-    def error_message(self, error_code, message=None, token=None):
-        """Create a JSON document that simulates the message served by
-        Overdrive given a certain error condition.
-        """
-        message = message or self.db.fresh_str()
-        token = token or self.db.fresh_str()
-        data = dict(errorCode=error_code, message=message, token=token)
-        return json.dumps(data)
-
-    def sample_json(self, filename):
-        data = self.data.sample_data(filename)
-        return data, json.loads(data)
-
-    def sync_patron_activity(self, patron: Patron):
-        self.api.sync_patron_activity(patron, "dummy pin")
-        return self.api.local_loans_and_holds(patron)
-
-
-@pytest.fixture(scope="function")
-def overdrive_api_fixture(
-    db: DatabaseTransactionFixture,
-    overdrive_files_fixture: OverdriveFilesFixture,
-) -> OverdriveAPIFixture:
-    return OverdriveAPIFixture(db, overdrive_files_fixture)
+    pass
 
 
 class TestOverdriveAPI:
@@ -253,28 +167,6 @@ class TestOverdriveAPI:
         # else the default will be used.
         l2 = transaction.library()
         assert "default" == fixture.api.ils_name(l2)
-
-    def test_make_link_safe(self):
-        # Unsafe characters are escaped.
-        assert "http://foo.com?q=%2B%3A%7B%7D" == OverdriveAPI.make_link_safe(
-            "http://foo.com?q=+:{}"
-        )
-
-        # Links to version 1 of the availability API are converted
-        # to links to version 2.
-        v1 = "https://qa.api.overdrive.com/v1/collections/abcde/products/12345/availability"
-        v2 = "https://qa.api.overdrive.com/v2/collections/abcde/products/12345/availability"
-        assert v2 == OverdriveAPI.make_link_safe(v1)
-
-        # We also handle the case of a trailing slash, just in case Overdrive
-        # starts serving links with trailing slashes.
-        v1 = v1 + "/"
-        v2 = v2 + "/"
-        assert v2 == OverdriveAPI.make_link_safe(v1)
-
-        # Links to other endpoints are not converted
-        leave_alone = "https://qa.api.overdrive.com/v1/collections/abcde/products/12345"
-        assert leave_alone == OverdriveAPI.make_link_safe(leave_alone)
 
     def test_hosts(self, overdrive_api_fixture: OverdriveAPIFixture):
         fixture = overdrive_api_fixture
@@ -2867,1300 +2759,3 @@ class TestSyncBookshelf:
         overdrive_api_fixture.sync_patron_activity(patron)
         assert 5 == len(patron.holds)
         assert overdrive_hold in patron.holds
-
-
-class TestOverdriveManifestFulfillment:
-    def test_response(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-
-        # An OverdriveManifestFulfillment just redirects the client
-        # directly to the manifest file
-        info = OverdriveManifestFulfillment(
-            "http://content-link/",
-            "scope string",
-            "access token",
-        )
-        response = info.response()
-        assert 302 == response.status_code
-        assert "" == response.get_data(as_text=True)
-        headers = response.headers
-        assert "text/plain" == headers["Content-Type"]
-
-        # These are the important headers; the location of the manifest file
-        # and the scope necessary to initiate Patron Authentication for
-        # it.
-        assert "scope string" == headers["X-Overdrive-Scope"]
-        assert "Bearer access token" == headers["X-Overdrive-Patron-Authorization"]
-        assert "http://content-link/" == headers["Location"]
-
-
-class TestOverdriveCirculationMonitor:
-    def test_run(self, overdrive_api_fixture: OverdriveAPIFixture, time_fixture: Time):
-        db = overdrive_api_fixture.db
-
-        # An end-to-end test verifying that this Monitor manages its
-        # state across multiple runs.
-        #
-        # This tests a lot of code that's technically not in Monitor,
-        # but when the Monitor API changes, it may require changes to
-        # this particular monitor, and it's good to have a test that
-        # will fail if that's true.
-        class Mock(OverdriveCirculationMonitor):
-            def catch_up_from(self, start, cutoff, progress):
-                self.catch_up_from_called_with = (start, cutoff, progress)
-
-        monitor = Mock(db.session, overdrive_api_fixture.collection)
-
-        monitor.run()
-        start, cutoff, progress = monitor.catch_up_from_called_with
-        now = utc_now()
-
-        # The first time this Monitor is called, its 'start time' is
-        # the current time, and we ask for an overlap of one minute.
-        # This isn't very effective, but we have to start somewhere.
-        #
-        # (This isn't how the Overdrive collection is initially
-        # populated, BTW -- that's NewTitlesOverdriveCollectionMonitor.)
-        time_fixture.time_eq(start, now - monitor.OVERLAP)
-        time_fixture.time_eq(cutoff, now)
-        timestamp = monitor.timestamp()
-        assert start == timestamp.start
-        assert cutoff == timestamp.finish
-
-        # The second time the Monitor is called, its 'start time'
-        # is one minute before the previous cutoff time.
-        monitor.run()
-        new_start, new_cutoff, new_progress = monitor.catch_up_from_called_with
-        now = utc_now()
-        assert new_start == cutoff - monitor.OVERLAP
-        time_fixture.time_eq(new_cutoff, now)
-
-    def test_catch_up_from(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-
-        # catch_up_from() asks Overdrive about recent changes by
-        # calling recently_changed_ids().
-        #
-        # It mirrors those changes locally by calling
-        # update_licensepool().
-        #
-        # If this is our first time encountering a book, a
-        # DISTRIBUTOR_TITLE_ADD analytics event is sent out.
-        #
-        # The method stops when should_stop() -- called on every book
-        # -- returns True.
-        class MockAPI:
-            def __init__(self, *ignore, **kwignore):
-                self.licensepools = []
-                self.update_licensepool_calls = []
-
-            def update_licensepool(self, book_id):
-                pool, is_new, is_changed = self.licensepools.pop(0)
-                self.update_licensepool_calls.append((book_id, pool))
-                return pool, is_new, is_changed
-
-        class MockAnalytics(Analytics):
-            def __init__(self):
-                self.events = []
-
-            def collect_event(self, *args):
-                self.events.append(args)
-
-        class MockMonitor(OverdriveCirculationMonitor):
-            recently_changed_ids_called_with = None
-            should_stop_calls = []
-
-            def recently_changed_ids(self, start, cutoff):
-                self.recently_changed_ids_called_with = (start, cutoff)
-                return [1, 2, None, 3, 4]
-
-            def should_stop(self, start, book, is_changed):
-                # We're going to stop after the third valid book,
-                # ensuring that we never ask 'Overdrive' for the
-                # fourth book.
-                self.should_stop_calls.append((start, book, is_changed))
-                if book == 3:
-                    return True
-                return False
-
-        monitor = MockMonitor(
-            db.session,
-            overdrive_api_fixture.collection,
-            api_class=MockAPI,  # type: ignore[arg-type]
-            analytics=MockAnalytics(),
-        )
-        api = cast(MockAPI, monitor.api)
-
-        # A MockAnalytics object was created and is ready to receive analytics
-        # events.
-        assert isinstance(monitor.analytics, MockAnalytics)
-
-        # The 'Overdrive API' is ready to tell us about four books,
-        # but only one of them (the first) represents a change from what
-        # we already know.
-        lp1 = db.licensepool(None)
-        lp1.last_checked = utc_now()
-        lp2 = db.licensepool(None)
-        lp3 = db.licensepool(None)
-        lp4 = MagicMock()
-        api.licensepools.append((lp1, True, True))
-        api.licensepools.append((lp2, False, False))
-        api.licensepools.append((lp3, False, True))
-        api.licensepools.append(lp4)
-
-        progress = TimestampData()
-        start = MagicMock()
-        cutoff = MagicMock()
-        monitor.catch_up_from(start, cutoff, progress)
-
-        # The monitor called recently_changed_ids with the start and
-        # cutoff times. It returned five 'books', one of which was None --
-        # simulating a lack of data from Overdrive.
-        assert (start, cutoff) == monitor.recently_changed_ids_called_with
-
-        # The monitor ignored the empty book and called
-        # update_licensepool on the first three valid 'books'. The
-        # mock API delivered the first three LicensePools from the
-        # queue.
-        assert [(1, lp1), (2, lp2), (3, lp3)] == api.update_licensepool_calls
-
-        # After each book was processed, should_stop was called, using
-        # the LicensePool, the start date, plus information about
-        # whether the LicensePool was changed (or created) during
-        # update_licensepool().
-        assert [
-            (start, 1, True),
-            (start, 2, False),
-            (start, 3, True),
-        ] == monitor.should_stop_calls
-
-        # should_stop returned True on the third call, and at that
-        # point we gave up.
-
-        # The fourth (bogus) LicensePool is still in api.licensepools,
-        # because we never asked for it.
-        assert [lp4] == api.licensepools
-
-        # A single analytics event was sent out, for the first LicensePool,
-        # the one that update_licensepool said was new.
-        #
-        # No more DISTRIBUTOR events
-        assert len(monitor.analytics.events) == 0
-
-        # The incoming TimestampData object was updated with
-        # a summary of what happened.
-        #
-        # We processed four books: 1, 2, None (which was ignored)
-        # and 3.
-        assert "Books processed: 4." == progress.achievements
-
-    def test_catch_up_from_with_failures_retried(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        """Check that book failures are retried."""
-        db = overdrive_api_fixture.db
-
-        class MockAPI:
-            tries: dict[str, int] = {}
-
-            def __init__(self, *ignore, **kwignore):
-                self.licensepools = []
-                self.update_licensepool_calls = []
-
-            def recently_changed_ids(self, start, cutoff):
-                return [1, 2, 3]
-
-            def update_licensepool(self, book_id):
-                current_count = self.tries.get(str(book_id)) or 0
-                current_count = current_count + 1
-                self.tries[str(book_id)] = current_count
-
-                if current_count < 1:
-                    raise StaleDataError("Ouch!")
-                elif current_count < 2:
-                    raise ObjectDeletedError({}, "Ouch Deleted!")
-
-                pool, is_new, is_changed = self.licensepools.pop(0)
-                self.update_licensepool_calls.append((book_id, pool))
-                return pool, is_new, is_changed
-
-        class MockAnalytics(Analytics):
-            def __init__(self):
-                self.events = []
-
-            def collect_event(self, *args):
-                self.events.append(args)
-
-        monitor = OverdriveCirculationMonitor(
-            db.session,
-            overdrive_api_fixture.collection,
-            api_class=MockAPI,  # type: ignore[arg-type]
-            analytics=MockAnalytics(),
-        )
-        api = cast(MockAPI, monitor.api)
-
-        # A MockAnalytics object was created and is ready to receive analytics
-        # events.
-        assert isinstance(monitor.analytics, MockAnalytics)
-
-        lp1 = db.licensepool(None)
-        lp1.last_checked = utc_now()
-        lp2 = db.licensepool(None)
-        lp3 = db.licensepool(None)
-        api.licensepools.append((lp1, True, True))
-        api.licensepools.append((lp2, False, False))
-        api.licensepools.append((lp3, False, True))
-
-        progress = TimestampData()
-        start = MagicMock()
-        cutoff = MagicMock()
-        monitor.catch_up_from(start, cutoff, progress)
-
-        assert api.tries["1"] == 2
-        assert api.tries["2"] == 2
-        assert api.tries["3"] == 2
-        assert not progress.is_failure
-
-    def test_catch_up_from_with_failures_all(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        """If an individual book fails, the import continues, but ends in failure after handling all the books."""
-        db = overdrive_api_fixture.db
-
-        class MockAPI:
-            tries: dict[str, int] = {}
-
-            def __init__(self, *ignore, **kwignore):
-                self.licensepools = []
-                self.update_licensepool_calls = []
-
-            def recently_changed_ids(self, start, cutoff):
-                return [1, 2, 3]
-
-            def update_licensepool(self, book_id):
-                current_count = self.tries.get(str(book_id)) or 0
-                current_count = current_count + 1
-                self.tries[str(book_id)] = current_count
-                raise Exception("Generic exception that will cause bypass retries")
-
-        class MockAnalytics(Analytics):
-            def __init__(self):
-                self.events = []
-
-            def collect_event(self, *args):
-                self.events.append(args)
-
-        monitor = OverdriveCirculationMonitor(
-            db.session,
-            overdrive_api_fixture.collection,
-            api_class=MockAPI,  # type: ignore[arg-type]
-            analytics=MockAnalytics(),
-        )
-
-        api = cast(MockAPI, monitor.api)
-
-        # A MockAnalytics object was created and is ready to receive analytics
-        # events.
-        assert isinstance(monitor.analytics, MockAnalytics)
-
-        lp1 = db.licensepool(None)
-        lp1.last_checked = utc_now()
-        lp2 = db.licensepool(None)
-        lp3 = db.licensepool(None)
-        api.licensepools.append((lp1, True, True))
-        api.licensepools.append((lp2, False, False))
-        api.licensepools.append((lp3, False, True))
-
-        progress = TimestampData()
-        start = MagicMock()
-        cutoff = MagicMock()
-        monitor.catch_up_from(start, cutoff, progress)
-
-        assert api.tries["1"] == 1
-        assert api.tries["2"] == 1
-        assert api.tries["3"] == 1
-        assert progress.is_failure
-
-    def test_retries_for_retryable_errors(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        """If  individual books fail due to retryable conditions, confirm success"""
-        db = overdrive_api_fixture.db
-        book1 = 1
-        book2 = 2
-
-        class MockAPI:
-            tries: dict[str, int] = {}
-
-            def __init__(self, *ignore, **kwignore):
-                self.licensepools = []
-                self.update_licensepool_calls = []
-
-            def recently_changed_ids(self, start, cutoff):
-                return [book1, book2]
-
-            def update_licensepool(self, book_id):
-                current_count = self.tries.get(str(book_id)) or 0
-                current_count = current_count + 1
-                self.tries[str(book_id)] = current_count
-                if book_id == 1:
-                    if current_count == 1:
-                        raise StaleDataError("stale data")
-                elif book_id == 2:
-                    if current_count == 1:
-                        raise ObjectDeletedError({}, "object deleted")
-
-                return None, None, False
-
-        monitor = OverdriveCirculationMonitor(
-            db.session,
-            overdrive_api_fixture.collection,
-            api_class=MockAPI,  # type: ignore[arg-type]
-        )
-
-        api = cast(MockAPI, monitor.api)
-
-        lp1 = db.licensepool(None)
-        lp1.last_checked = utc_now()
-        lp2 = db.licensepool(None)
-        api.licensepools.append((lp1, True, True))
-        api.licensepools.append((lp2, False, False))
-
-        progress = TimestampData()
-        start = MagicMock()
-        cutoff = MagicMock()
-        monitor.catch_up_from(start, cutoff, progress)
-
-        for b in [book1, book2]:
-            assert api.tries[str(b)] == 2
-        assert not progress.is_failure
-
-
-class TestNewTitlesOverdriveCollectionMonitor:
-    def test_recently_changed_ids(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-
-        class MockAPI:
-            def __init__(self, *args, **kwargs):
-                pass
-
-            def all_ids(self):
-                return "all of the ids"
-
-        monitor = NewTitlesOverdriveCollectionMonitor(
-            db.session, overdrive_api_fixture.collection, api_class=MockAPI
-        )
-        assert "all of the ids" == monitor.recently_changed_ids(
-            MagicMock(), MagicMock()
-        )
-
-    def test_should_stop(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-        monitor = NewTitlesOverdriveCollectionMonitor(
-            db.session, overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-        )
-
-        # for this test, we will not count consecutive out of scope dates: if one
-        # title is out of scope, we should stop.
-        NewTitlesOverdriveCollectionMonitor.MAX_CONSECUTIVE_OUT_OF_SCOPE_DATES = 0
-
-        m = monitor.should_stop
-
-        # If the monitor has never run before, we need to keep going
-        # until we run out of books.
-        assert False == m(None, MagicMock(), MagicMock())
-        assert False == m(monitor.NEVER, MagicMock(), MagicMock())  # type: ignore[arg-type]
-
-        # If information is missing or invalid, we assume that we
-        # should keep going.
-        start = datetime_utc(2018, 1, 1)
-        assert False == m(start, {}, MagicMock())
-        assert False == m(start, {"date_added": None}, MagicMock())
-        assert False == m(start, {"date_added": "Not a date"}, MagicMock())
-
-        # Here, we're actually comparing real dates, using the date
-        # format found in the Overdrive API. A date that's after the
-        # `start` date means we should keep going backwards. A date before
-        # the `start` date means we should stop.
-        assert False == m(
-            start, {"date_added": "2019-07-12T11:06:38.157+01:00"}, MagicMock()
-        )
-        assert True == m(
-            start, {"date_added": "2017-07-12T11:06:38.157-04:00"}, MagicMock()
-        )
-
-    def test_should_stop_with_consecutive_data_threshold_gt_zero(
-        self, overdrive_api_fixture: OverdriveAPIFixture, caplog
-    ):
-        caplog.set_level(logging.INFO)
-
-        db = overdrive_api_fixture.db
-        monitor = NewTitlesOverdriveCollectionMonitor(
-            db.session, overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-        )
-
-        # for this test, we will count consecutive out of scope date
-        NewTitlesOverdriveCollectionMonitor.MAX_CONSECUTIVE_OUT_OF_SCOPE_DATES = 1
-
-        m = monitor.should_stop
-
-        start = datetime_utc(2018, 1, 1)
-
-        # in scope - should continue
-        in_scope_properties = {"date_added": "2019-07-12T11:06:38.157+01:00"}
-        assert False == m(start, in_scope_properties, MagicMock())
-
-        assert "Date added: 2019-07-12 11:06:38.157000+01:00" in caplog.messages[-1]
-
-        # out of scope but counter threshold not yet exceeded: should continue
-        out_of_scope_properties = {"date_added": "2017-07-12T11:06:38.157-04:00"}
-        assert not m(start, out_of_scope_properties, MagicMock())
-
-        assert "Date added: 2017-07-12 11:06:38.157000-04:00" in caplog.messages[-1]
-
-        # in scope - should continue, expect reset
-        assert not m(start, in_scope_properties, MagicMock())
-
-        assert (
-            "We encountered a title that was added within our scope that "
-            "followed a title that was out of scope"
-        ) in caplog.messages[-1]
-
-        # out of scope but counter threshold not yet exceeded: should continue
-        assert not m(start, out_of_scope_properties, MagicMock())
-
-        # second out of scope:  threshold exceeded:  should stop
-        assert m(start, out_of_scope_properties, MagicMock())
-
-        assert (
-            "Max consecutive out of scope date threshold of 1 breached!"
-            in caplog.messages[-1]
-        )
-
-
-class TestNewTitlesOverdriveCollectionMonitor2:
-    def test_should_stop(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-        monitor = RecentOverdriveCollectionMonitor(
-            db.session, overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-        )
-        assert 0 == monitor.consecutive_unchanged_books
-        m = monitor.should_stop
-
-        # This book hasn't been changed, but we're under the limit, so we should
-        # keep going.
-        assert False == m(MagicMock(), MagicMock(), False)
-        assert 1 == monitor.consecutive_unchanged_books
-
-        assert False == m(MagicMock(), MagicMock(), False)
-        assert 2 == monitor.consecutive_unchanged_books
-
-        # This book has changed, so our counter gets reset.
-        assert False == m(MagicMock(), MagicMock(), True)
-        assert 0 == monitor.consecutive_unchanged_books
-
-        # When we're at the limit, and another book comes along that hasn't
-        # been changed, _then_ we decide to stop.
-        monitor.consecutive_unchanged_books = (
-            monitor.MAXIMUM_CONSECUTIVE_UNCHANGED_BOOKS
-        )
-        assert True == m(MagicMock(), MagicMock(), False)
-        assert (
-            monitor.MAXIMUM_CONSECUTIVE_UNCHANGED_BOOKS + 1
-            == monitor.consecutive_unchanged_books
-        )
-
-
-class TestOverdriveFormatSweep:
-    def test_process_item(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-        # Validate the standard CollectionMonitor interface.
-        monitor = OverdriveFormatSweep(
-            db.session, overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-        )
-        mock_api = cast(MockOverdriveAPI, monitor.api)
-        mock_api.queue_collection_token()
-        # We're not testing that the work actually gets done (that's
-        # tested in test_update_formats), only that the monitor
-        # implements the expected process_item API without crashing.
-        mock_api.queue_response(404)
-        edition, pool = db.edition(with_license_pool=True)
-        monitor.process_item(pool.identifier)
-
-    def test_process_item_multiple_licence_pools(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        db = overdrive_api_fixture.db
-        # Make sure that we only call update_formats once when an item
-        # is part of multiple licensepools.
-
-        class MockApi(MockOverdriveAPI):
-            update_format_calls = 0
-
-            def update_formats(self, licensepool):
-                self.update_format_calls += 1
-
-        monitor = OverdriveFormatSweep(
-            db.session, overdrive_api_fixture.collection, api_class=MockApi
-        )
-        mock_api = cast(MockApi, monitor.api)
-        mock_api.queue_collection_token()
-        mock_api.queue_response(404)
-
-        edition = db.edition()
-        collection1 = db.collection(name="Collection 1")
-        pool1 = db.licensepool(edition, collection=collection1)
-
-        collection2 = db.collection(name="Collection 2")
-        pool2 = db.licensepool(edition, collection=collection2)
-
-        monitor.process_item(pool1.identifier)
-        assert 1 == mock_api.update_format_calls
-
-
-class TestReaper:
-    def test_instantiate(self, overdrive_api_fixture: OverdriveAPIFixture):
-        db = overdrive_api_fixture.db
-        # Validate the standard CollectionMonitor interface.
-        monitor = OverdriveCollectionReaper(
-            db.session, overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-        )
-
-
-class TestOverdriveRepresentationExtractor:
-    def test_availability_info(self, overdrive_api_fixture: OverdriveAPIFixture):
-        data, raw = overdrive_api_fixture.sample_json("overdrive_book_list.json")
-        availability = OverdriveRepresentationExtractor.availability_link_list(raw)
-        # Every item in the list has a few important values.
-        for item in availability:
-            for key in "availability_link", "author_name", "id", "title", "date_added":
-                assert key in item
-
-        # Also run a spot check on the actual values.
-        spot = availability[0]
-        assert "210bdcad-29b7-445f-8d05-cdbb40abc03a" == spot["id"]
-        assert "King and Maxwell" == spot["title"]
-        assert "David Baldacci" == spot["author_name"]
-        assert "2013-11-12T14:13:00-05:00" == spot["date_added"]
-
-    def test_availability_info_missing_data(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        # overdrive_book_list_missing_data.json has two products. One
-        # only has a title, the other only has an ID.
-        data, raw = overdrive_api_fixture.sample_json(
-            "overdrive_book_list_missing_data.json"
-        )
-        [item] = OverdriveRepresentationExtractor.availability_link_list(raw)
-
-        # We got a data structure -- full of missing data -- for the
-        # item that has an ID.
-        assert "i only have an id" == item["id"]
-        assert None == item["title"]
-        assert None == item["author_name"]
-        assert None == item["date_added"]
-
-        # We did not get a data structure for the item that only has a
-        # title, because an ID is required -- otherwise we don't know
-        # what book we're talking about.
-
-    def test_link(self, overdrive_api_fixture: OverdriveAPIFixture):
-        data, raw = overdrive_api_fixture.sample_json("overdrive_book_list.json")
-        expect = OverdriveAPI.make_link_safe(
-            "http://api.overdrive.com/v1/collections/collection-id/products?limit=300&offset=0&lastupdatetime=2014-04-28%2009:25:09&sort=popularity:desc&formats=ebook-epub-open,ebook-epub-adobe,ebook-pdf-adobe,ebook-pdf-open"
-        )
-        assert expect == OverdriveRepresentationExtractor.link(raw, "first")
-
-    def test_book_info_to_circulation(self, overdrive_api_fixture: OverdriveAPIFixture):
-        # Tests that can convert an overdrive json block into a CirculationData object.
-        fixture = overdrive_api_fixture
-        session = overdrive_api_fixture.db.session
-
-        raw, info = fixture.sample_json("overdrive_availability_information_2.json")
-        extractor = OverdriveRepresentationExtractor(fixture.api)
-        circulationdata = extractor.book_info_to_circulation(info)
-
-        # NOTE: It's not realistic for licenses_available and
-        # patrons_in_hold_queue to both be nonzero; this is just to
-        # verify that the test picks up whatever data is in the
-        # document.
-        assert 3 == circulationdata.licenses_owned
-        assert 1 == circulationdata.licenses_available
-        assert 10 == circulationdata.patrons_in_hold_queue
-
-        # Related IDs.
-        identifier = circulationdata.primary_identifier(session)
-        assert (Identifier.OVERDRIVE_ID, "2a005d55-a417-4053-b90d-7a38ca6d2065") == (
-            identifier.type,
-            identifier.identifier,
-        )
-
-    def test_book_info_to_circulation_advantage(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        # Overdrive Advantage accounts (a.k.a. "child" or "sub" accounts derive
-        # different information from the same API responses as "main" Overdrive
-        # accounts.
-        fixture = overdrive_api_fixture
-        raw, info = fixture.sample_json("overdrive_availability_advantage.json")
-
-        extractor = OverdriveRepresentationExtractor(fixture.api)
-        # Calling in the context of a main account should return a count of
-        # the main account and any shared sub account owned and available.
-        consortial_data = extractor.book_info_to_circulation(info)
-        assert 10 == consortial_data.licenses_owned
-        assert 10 == consortial_data.licenses_available
-
-        # Pretend to be an API for an Overdrive Advantage collection with
-        # library ID 61.
-        extractor = OverdriveRepresentationExtractor(MagicMock(advantage_library_id=61))
-        advantage_data = extractor.book_info_to_circulation(info)
-        assert 1 == advantage_data.licenses_owned
-        assert 1 == advantage_data.licenses_available
-
-        # Both collections have the same information about active
-        # holds, because that information is not split out by
-        # collection.
-        assert 0 == advantage_data.patrons_in_hold_queue
-        assert 0 == consortial_data.patrons_in_hold_queue
-
-        # If for whatever reason Overdrive doesn't mention the
-        # relevant collection at all, no collection-specific
-        # information is gleaned.
-        #
-        # TODO: It would probably be better not to return a
-        # CirculationData object at all, but this shouldn't happen in
-        # a real scenario.
-        extractor = OverdriveRepresentationExtractor(MagicMock(advantage_library_id=62))
-        advantage_data = extractor.book_info_to_circulation(info)
-        assert 0 == advantage_data.licenses_owned
-        assert 0 == advantage_data.licenses_available
-
-        # Pretend to be an API for an Overdrive Advantage collection with
-        # library ID 63 which contains shared copies.
-        extractor = OverdriveRepresentationExtractor(MagicMock(advantage_library_id=63))
-        advantage_data = extractor.book_info_to_circulation(info)
-        # since these copies are shared and counted as part of the main
-        # context we do not count them here.
-        assert 0 == advantage_data.licenses_owned
-        assert 0 == advantage_data.licenses_available
-
-    def test_not_found_error_to_circulationdata(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        fixture = overdrive_api_fixture
-        transaction = fixture.db
-        raw, info = fixture.sample_json("overdrive_availability_not_found.json")
-
-        # By default, a "NotFound" error can't be converted to a
-        # CirculationData object, because we don't know _which_ book it
-        # was that wasn't found.
-        extractor = OverdriveRepresentationExtractor(fixture.api)
-        m = extractor.book_info_to_circulation
-        with pytest.raises(
-            PalaceValueError, match="Book must have an id to be processed"
-        ):
-            m(info)
-
-        # However, if an ID was added to `info` ahead of time (as the
-        # circulation code does), we do know, and we can create a
-        # CirculationData.
-        identifier = transaction.identifier(identifier_type=Identifier.OVERDRIVE_ID)
-        info["id"] = identifier.identifier
-        data = m(info)
-        assert identifier == data.primary_identifier(transaction.session)
-        assert 0 == data.licenses_owned
-        assert 0 == data.licenses_available
-        assert 0 == data.patrons_in_hold_queue
-
-    def test_book_info_with_metadata(self, overdrive_api_fixture: OverdriveAPIFixture):
-        # Tests that can convert an overdrive json block into a Metadata object.
-
-        raw, info = overdrive_api_fixture.sample_json("overdrive_metadata.json")
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-
-        assert "Agile Documentation" == metadata.title
-        assert (
-            "Agile Documentation A Pattern Guide to Producing Lightweight Documents for Software Projects"
-            == metadata.sort_title
-        )
-        assert (
-            "A Pattern Guide to Producing Lightweight Documents for Software Projects"
-            == metadata.subtitle
-        )
-        assert Edition.BOOK_MEDIUM == metadata.medium
-        assert "Wiley Software Patterns" == metadata.series
-        assert "eng" == metadata.language
-        assert "Wiley" == metadata.publisher
-        assert "John Wiley & Sons, Inc." == metadata.imprint
-        assert 2005 == metadata.published.year
-        assert 1 == metadata.published.month
-        assert 31 == metadata.published.day
-
-        [author] = metadata.contributors
-        assert "Rüping, Andreas" == author.sort_name
-        assert "Andreas R&#252;ping" == author.display_name
-        assert [Contributor.Role.AUTHOR] == author.roles
-
-        subjects = sorted(metadata.subjects, key=lambda x: x.identifier)
-
-        assert [
-            ("Computer Technology", Subject.OVERDRIVE, 100),
-            ("Nonfiction", Subject.OVERDRIVE, 100),
-            ("Object Technologies - Miscellaneous", "tag", 1),
-        ] == [(x.identifier, x.type, x.weight) for x in subjects]
-
-        # Related IDs.
-        assert (Identifier.OVERDRIVE_ID, "3896665d-9d81-4cac-bd43-ffc5066de1f5") == (
-            metadata.primary_identifier.type,
-            metadata.primary_identifier.identifier,
-        )
-
-        ids = [(x.type, x.identifier) for x in metadata.identifiers]
-
-        # The original data contains an actual ASIN and ISBN, plus a blank
-        # ASIN and three invalid ISBNs: one which is common placeholder
-        # text, one which is mis-typed and has a bad check digit, and one
-        # which has an invalid character; the bad identifiers do not show
-        # up here.
-        assert [
-            (Identifier.ASIN, "B000VI88N2"),
-            (Identifier.ISBN, "9780470856246"),
-            (Identifier.OVERDRIVE_ID, "3896665d-9d81-4cac-bd43-ffc5066de1f5"),
-        ] == sorted(ids)
-
-        # Available formats.
-        [kindle, pdf] = sorted(
-            metadata.circulation.formats, key=lambda x: x.content_type
-        )
-        assert DeliveryMechanism.KINDLE_CONTENT_TYPE == kindle.content_type
-        assert DeliveryMechanism.KINDLE_DRM == kindle.drm_scheme
-
-        assert Representation.PDF_MEDIA_TYPE == pdf.content_type
-        assert DeliveryMechanism.ADOBE_DRM == pdf.drm_scheme
-
-        # Links to various resources.
-        shortd, image, longd = sorted(metadata.links, key=lambda x: x.rel)
-
-        assert Hyperlink.DESCRIPTION == longd.rel
-        assert longd.content.startswith("<p>Software documentation")
-
-        assert Hyperlink.SHORT_DESCRIPTION == shortd.rel
-        assert shortd.content.startswith("<p>Software documentation")
-        assert len(shortd.content) < len(longd.content)
-
-        assert Hyperlink.IMAGE == image.rel
-        assert (
-            "http://images.contentreserve.com/ImageType-100/0128-1/%7B3896665D-9D81-4CAC-BD43-FFC5066DE1F5%7DImg100.jpg"
-            == image.href
-        )
-
-        thumbnail = image.thumbnail
-
-        assert Hyperlink.THUMBNAIL_IMAGE == thumbnail.rel
-        assert (
-            "http://images.contentreserve.com/ImageType-200/0128-1/%7B3896665D-9D81-4CAC-BD43-FFC5066DE1F5%7DImg200.jpg"
-            == thumbnail.href
-        )
-
-        # Measurements associated with the book.
-
-        measurements = metadata.measurements
-        popularity = [
-            x for x in measurements if x.quantity_measured == Measurement.POPULARITY
-        ][0]
-        assert 2 == popularity.value
-
-        rating = [x for x in measurements if x.quantity_measured == Measurement.RATING][
-            0
-        ]
-        assert 1 == rating.value
-
-        # Request only the bibliographic information.
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(
-            info, include_bibliographic=True, include_formats=False
-        )
-
-        assert "Agile Documentation" == metadata.title
-        assert None == metadata.circulation
-
-        # Request only the format information.
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(
-            info, include_bibliographic=False, include_formats=True
-        )
-
-        assert None == metadata.title
-
-        [kindle, pdf] = sorted(
-            metadata.circulation.formats, key=lambda x: x.content_type
-        )
-        assert DeliveryMechanism.KINDLE_CONTENT_TYPE == kindle.content_type
-        assert DeliveryMechanism.KINDLE_DRM == kindle.drm_scheme
-
-        assert Representation.PDF_MEDIA_TYPE == pdf.content_type
-        assert DeliveryMechanism.ADOBE_DRM == pdf.drm_scheme
-
-    def test_audiobook_info(self, overdrive_api_fixture: OverdriveAPIFixture):
-        # This book will be available in three formats: a link to the
-        # Overdrive Read website, a manifest file that SimplyE can
-        # download, and the legacy format used by the mobile app
-        # called 'Overdrive'.
-        raw, info = overdrive_api_fixture.sample_json("audiobook.json")
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-        streaming, manifest, legacy = sorted(
-            metadata.circulation.formats, key=lambda x: x.content_type
-        )
-        assert DeliveryMechanism.STREAMING_AUDIO_CONTENT_TYPE == streaming.content_type
-        assert (
-            MediaTypes.OVERDRIVE_AUDIOBOOK_MANIFEST_MEDIA_TYPE == manifest.content_type
-        )
-        assert "application/x-od-media" == legacy.content_type
-        assert (
-            metadata.duration == 10 * 3600 + 9 * 60 + 1
-        )  # The last formats' duration attribute
-
-        # The last format will be invalid, so only the first format should work
-        info["formats"][1]["duration"] = "10:09"  # Invalid format
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-        assert (
-            metadata.duration == 10 * 3600 + 9 * 60 + 0
-        )  # The first formats' duration attribute
-
-    def test_book_info_with_sample(self, overdrive_api_fixture: OverdriveAPIFixture):
-        # This book has two samples; one available as a direct download and
-        # one available through a manifest file.
-        raw, info = overdrive_api_fixture.sample_json("has_sample.json")
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-        samples = [x for x in metadata.links if x.rel == Hyperlink.SAMPLE]
-        epub_sample, manifest_sample = sorted(samples, key=lambda x: x.media_type or "")
-
-        # Here's the direct download.
-        assert (
-            "http://excerpts.contentreserve.com/FormatType-410/1071-1/9BD/24F/82/BridesofConvenienceBundle9781426803697.epub"
-            == epub_sample.href
-        )
-        assert MediaTypes.EPUB_MEDIA_TYPE == epub_sample.media_type
-
-        # Here's the manifest.
-        assert (
-            "https://samples.overdrive.com/?crid=9BD24F82-35C0-4E0A-B5E7-BCFED07835CF&.epub-sample.overdrive.com"
-            == manifest_sample.href
-        )
-        # Assert we have the end content type of the sample, no DRM formats
-        assert "text/html" == manifest_sample.media_type
-
-    def test_book_info_with_unknown_sample(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        raw, info = overdrive_api_fixture.sample_json("has_sample.json")
-
-        # Just use one format, and change a sample type to unknown
-        # Only one (known sample) should be extracted then
-        info["formats"] = [info["formats"][1]]
-        info["formats"][0]["samples"][1]["formatType"] = "overdrive-unknown"
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-        samples = [x for x in metadata.links if x.rel == Hyperlink.SAMPLE]
-
-        assert 1 == len(samples)
-        assert samples[0].media_type == MediaTypes.EPUB_MEDIA_TYPE
-
-    def test_book_info_with_grade_levels(
-        self, overdrive_api_fixture: OverdriveAPIFixture
-    ):
-        raw, info = overdrive_api_fixture.sample_json("has_grade_levels.json")
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-
-        grade_levels = sorted(
-            x.identifier or "fail"
-            for x in metadata.subjects
-            if x.type == Subject.GRADE_LEVEL
-        )
-        assert ["Grade 4", "Grade 5", "Grade 6", "Grade 7", "Grade 8"] == grade_levels
-
-    def test_book_info_with_awards(self, overdrive_api_fixture: OverdriveAPIFixture):
-        raw, info = overdrive_api_fixture.sample_json("has_awards.json")
-        metadata = OverdriveRepresentationExtractor.book_info_to_metadata(info)
-
-        [awards] = [
-            x
-            for x in metadata.measurements
-            if Measurement.AWARDS == x.quantity_measured
-        ]
-        assert 1 == awards.value
-        assert 1 == awards.weight
-
-    def test_image_link_to_linkdata(self):
-        def m(link):
-            return OverdriveRepresentationExtractor.image_link_to_linkdata(link, "rel")
-
-        # Test missing data.
-        assert None == m(None)
-        assert None == m(dict())
-
-        # Test an ordinary success case.
-        url = "http://images.overdrive.com/image.png"
-        type = "image/type"
-        data = m(dict(href=url, type=type))
-        assert isinstance(data, LinkData)
-        assert url == data.href
-        assert type == data.media_type
-
-        # Test a case where no media type is provided.
-        data = m(dict(href=url))
-        assert None == data.media_type
-
-        # Verify that invalid URLs are made link-safe.
-        data = m(dict(href="http://api.overdrive.com/v1/foo:bar"))
-        assert "http://api.overdrive.com/v1/foo%3Abar" == data.href
-
-        # Stand-in cover images are detected and filtered out.
-        data = m(
-            dict(
-                href="https://img1.od-cdn.com/ImageType-100/0293-1/{00000000-0000-0000-0000-000000000002}Img100.jpg"
-            )
-        )
-        assert None == data
-
-    def test_internal_formats(self):
-        # Overdrive's internal format names may correspond to one or more
-        # delivery mechanisms.
-        def assert_formats(overdrive_name, *expect):
-            actual = OverdriveRepresentationExtractor.internal_formats(overdrive_name)
-            assert list(expect) == list(actual)
-
-        # Most formats correspond to one delivery mechanism.
-        assert_formats(
-            "ebook-pdf-adobe", (MediaTypes.PDF_MEDIA_TYPE, DeliveryMechanism.ADOBE_DRM)
-        )
-
-        assert_formats(
-            "ebook-epub-open", (MediaTypes.EPUB_MEDIA_TYPE, DeliveryMechanism.NO_DRM)
-        )
-
-        # ebook-overdrive and audiobook-overdrive each correspond to
-        # two delivery mechanisms.
-        assert_formats(
-            "ebook-overdrive",
-            (
-                MediaTypes.OVERDRIVE_EBOOK_MANIFEST_MEDIA_TYPE,
-                DeliveryMechanism.LIBBY_DRM,
-            ),
-            (
-                DeliveryMechanism.STREAMING_TEXT_CONTENT_TYPE,
-                DeliveryMechanism.STREAMING_DRM,
-            ),
-        )
-
-        assert_formats(
-            "audiobook-overdrive",
-            (
-                MediaTypes.OVERDRIVE_AUDIOBOOK_MANIFEST_MEDIA_TYPE,
-                DeliveryMechanism.LIBBY_DRM,
-            ),
-            (
-                DeliveryMechanism.STREAMING_AUDIO_CONTENT_TYPE,
-                DeliveryMechanism.STREAMING_DRM,
-            ),
-        )
-
-        # An unrecognized format does not correspond to any delivery
-        # mechanisms.
-        assert_formats("no-such-format")
-
-
-class TestOverdriveAdvantageAccount:
-    def test_no_advantage_accounts(self, overdrive_api_fixture: OverdriveAPIFixture):
-        """When there are no Advantage accounts, get_advantage_accounts()
-        returns an empty list.
-        """
-        fixture = overdrive_api_fixture
-        fixture.api.queue_collection_token()
-        assert [] == list(fixture.api.get_advantage_accounts())
-
-    def test_from_representation(self, overdrive_api_fixture: OverdriveAPIFixture):
-        """Test the creation of OverdriveAdvantageAccount objects
-        from Overdrive's representation of a list of accounts.
-        """
-        fixture = overdrive_api_fixture
-        raw, data = fixture.sample_json("advantage_accounts.json")
-        [ac1, ac2] = OverdriveAdvantageAccount.from_representation(raw)
-
-        # The two Advantage accounts have the same parent library ID.
-        assert "1225" == ac1.parent_library_id
-        assert "1225" == ac2.parent_library_id
-
-        # But they have different names and library IDs.
-        assert "3" == ac1.library_id
-        assert "The Other Side of Town Library" == ac1.name
-
-        assert "9" == ac2.library_id
-        assert "The Common Community Library" == ac2.name
-
-    def test_to_collection(self, overdrive_api_fixture: OverdriveAPIFixture):
-        # Test that we can turn an OverdriveAdvantageAccount object into
-        # a Collection object.
-        fixture = overdrive_api_fixture
-        transaction, session = (
-            fixture.db,
-            fixture.db.session,
-        )
-
-        account = OverdriveAdvantageAccount(
-            "parent_id",
-            "child_id",
-            "Library Name",
-            "token value",
-        )
-
-        # We can't just create a Collection object for this object because
-        # the parent doesn't exist.
-        with pytest.raises(ValueError) as excinfo:
-            account.to_collection(session)
-        assert "Cannot create a Collection whose parent does not already exist." in str(
-            excinfo.value
-        )
-
-        # So, create a Collection to be the parent.
-        parent = transaction.collection(
-            name="Parent",
-            protocol=OverdriveAPI,
-            settings=transaction.overdrive_settings(external_account_id="parent_id"),
-        )
-
-        # Now it works.
-        p, collection = account.to_collection(session)
-        assert p == parent
-        assert parent == collection.parent
-        assert (
-            collection.integration_configuration.settings_dict["external_account_id"]
-            == account.library_id
-        )
-        assert OverdriveAPI.label() == collection.protocol
-        assert Goals.LICENSE_GOAL == collection.integration_configuration.goal
-
-        # To ensure uniqueness, the collection was named after its
-        # parent.
-        assert f"{parent.name} / {account.name}" == collection.name
-
-
-@dataclass
-class OverdriveBibliographicCoverageProviderFixture:
-    overdrive: OverdriveAPIFixture
-    provider: OverdriveBibliographicCoverageProvider
-    api: MockOverdriveAPI
-
-
-@pytest.fixture
-def overdrive_biblio_provider_fixture(
-    overdrive_api_fixture: OverdriveAPIFixture,
-) -> OverdriveBibliographicCoverageProviderFixture:
-    overdrive = overdrive_api_fixture
-    provider = OverdriveBibliographicCoverageProvider(
-        overdrive_api_fixture.collection, api_class=MockOverdriveAPI
-    )
-    api = cast(MockOverdriveAPI, provider.api)
-    return OverdriveBibliographicCoverageProviderFixture(overdrive, provider, api)
-
-
-class TestOverdriveBibliographicCoverageProvider:
-    """Test the code that looks up bibliographic information from Overdrive."""
-
-    def test_script_instantiation(
-        self,
-        overdrive_biblio_provider_fixture: OverdriveBibliographicCoverageProviderFixture,
-    ):
-        """Test that RunCoverageProviderScript can instantiate
-        the coverage provider.
-        """
-
-        fixture = overdrive_biblio_provider_fixture
-        db = fixture.overdrive.db
-
-        script = RunCollectionCoverageProviderScript(
-            OverdriveBibliographicCoverageProvider,
-            db.session,
-            api_class=MockOverdriveAPI,
-        )
-        [provider] = script.providers
-        assert isinstance(provider, OverdriveBibliographicCoverageProvider)
-        assert isinstance(provider.api, MockOverdriveAPI)
-        assert fixture.overdrive.collection == provider.collection
-
-    def test_invalid_or_unrecognized_guid(
-        self,
-        overdrive_biblio_provider_fixture: OverdriveBibliographicCoverageProviderFixture,
-    ):
-        """A bad or malformed GUID can't get coverage."""
-        fixture = overdrive_biblio_provider_fixture
-        db = fixture.overdrive.db
-
-        identifier = db.identifier()
-        identifier.identifier = "bad guid"
-
-        error = '{"errorCode": "InvalidGuid", "message": "An invalid guid was given.", "token": "7aebce0e-2e88-41b3-b6d3-82bf15f8e1a2"}'
-        fixture.api.queue_response(200, content=error)
-
-        failure = fixture.provider.process_item(identifier)
-        assert isinstance(failure, CoverageFailure)
-        assert False == failure.transient
-        assert "Invalid Overdrive ID: bad guid" == failure.exception
-
-        # This is for when the GUID is well-formed but doesn't
-        # correspond to any real Overdrive book.
-        error = '{"errorCode": "NotFound", "message": "Not found in Overdrive collection.", "token": "7aebce0e-2e88-41b3-b6d3-82bf15f8e1a2"}'
-        fixture.api.queue_response(200, content=error)
-
-        failure = fixture.provider.process_item(identifier)
-        assert isinstance(failure, CoverageFailure)
-        assert False == failure.transient
-        assert "ID not recognized by Overdrive: bad guid" == failure.exception
-
-    def test_process_item_creates_presentation_ready_work(
-        self,
-        overdrive_biblio_provider_fixture: OverdriveBibliographicCoverageProviderFixture,
-    ):
-        """Test the normal workflow where we ask Overdrive for data,
-        Overdrive provides it, and we create a presentation-ready work.
-        """
-        fixture = overdrive_biblio_provider_fixture
-        db = fixture.overdrive.db
-
-        # Here's the book mentioned in overdrive_metadata.json.
-        identifier = db.identifier(identifier_type=Identifier.OVERDRIVE_ID)
-        identifier.identifier = "3896665d-9d81-4cac-bd43-ffc5066de1f5"
-
-        # This book has no LicensePool.
-        assert [] == identifier.licensed_through
-
-        # Run it through the OverdriveBibliographicCoverageProvider
-        raw, info = fixture.overdrive.sample_json("overdrive_metadata.json")
-        fixture.api.queue_response(200, content=raw)
-
-        [result] = fixture.provider.process_batch([identifier])
-        assert identifier == result
-
-        # A LicensePool was created, not because we know anything
-        # about how we've licensed this book, but to have a place to
-        # store the information about what formats the book is
-        # available in.
-        [pool] = identifier.licensed_through
-        assert 0 == pool.licenses_owned
-        [lpdm1, lpdm2] = pool.delivery_mechanisms
-        names = [x.delivery_mechanism.name for x in pool.delivery_mechanisms]
-        assert sorted(
-            [
-                "application/pdf (application/vnd.adobe.adept+xml)",
-                "Kindle via Amazon (Kindle DRM)",
-            ]
-        ) == sorted(names)
-
-        # A Work was created and made presentation ready.
-        assert "Agile Documentation" == pool.work.title
-        assert True == pool.work.presentation_ready
-
-
-class TestGenerateOverdriveAdvantageAccountList:
-    def test_generate_od_advantage_account_list(self, db: DatabaseTransactionFixture):
-        output_file_path = "test-output.csv"
-        circ_manager_name = "circ_man_name"
-        parent_library_name = "Parent"
-        parent_od_library_id = "parent_id"
-        child1_library_name = "child1"
-        child1_advantage_library_id = "1"
-        child1_token = "token1"
-        child2_library_name = "child2"
-        child2_advantage_library_id = "2"
-        child2_token = "token2"
-        client_key = "ck"
-        client_secret = "cs"
-        library_token = "lt"
-
-        library = db.library()
-        parent: Collection = MockOverdriveAPI.mock_collection(
-            db.session,
-            library,
-            name=parent_library_name,
-            library_id=parent_od_library_id,
-            client_key=client_key,
-            client_secret=client_secret,
-        )
-        child1: Collection = MockOverdriveAPI.mock_collection(
-            db.session,
-            library,
-            name=child1_library_name,
-            library_id=child1_advantage_library_id,
-        )
-        child1.parent = parent
-        overdrive_api = MockOverdriveAPI(db.session, parent)
-        mock_get_advantage_accounts = MagicMock()
-        mock_get_advantage_accounts.return_value = [
-            OverdriveAdvantageAccount(
-                parent_od_library_id,
-                child1_advantage_library_id,
-                child1_library_name,
-                child1_token,
-            ),
-            OverdriveAdvantageAccount(
-                parent_od_library_id,
-                child2_advantage_library_id,
-                child2_library_name,
-                child2_token,
-            ),
-        ]
-        overdrive_api.get_advantage_accounts = mock_get_advantage_accounts
-        overdrive_api._collection_token = library_token
-
-        with patch(
-            "palace.manager.api.overdrive.GenerateOverdriveAdvantageAccountList._create_overdrive_api"
-        ) as create_od_api:
-            create_od_api.return_value = overdrive_api
-            GenerateOverdriveAdvantageAccountList(db.session).do_run(
-                cmd_args=[
-                    "--output-file-path",
-                    output_file_path,
-                    "--circulation-manager-name",
-                    circ_manager_name,
-                ]
-            )
-
-            with open(output_file_path, newline="") as csv_file:
-                csvreader = csv.reader(csv_file)
-                for index, row in enumerate(csvreader):
-                    if index == 0:
-                        assert "cm" == row[0]
-                        assert "collection" == row[1]
-                        assert "overdrive_library_id" == row[2]
-                        assert "client_key" == row[3]
-                        assert "client_secret" == row[4]
-                        assert "library_token" == row[5]
-                        assert "advantage_name" == row[6]
-                        assert "advantage_id" == row[7]
-                        assert "advantage_token" == row[8]
-                        assert "already_configured" == row[9]
-                    elif index == 1:
-                        assert circ_manager_name == row[0]
-                        assert parent_library_name == row[1]
-                        assert parent_od_library_id == row[2]
-                        assert client_key == row[3]
-                        assert client_secret == row[4]
-                        assert library_token == row[5]
-                        assert child1_library_name == row[6]
-                        assert child1_advantage_library_id == row[7]
-                        assert child1_token == row[8]
-                        assert "True" == row[9]
-                    else:
-                        assert circ_manager_name == row[0]
-                        assert parent_library_name == row[1]
-                        assert parent_od_library_id == row[2]
-                        assert client_key == row[3]
-                        assert client_secret == row[4]
-                        assert library_token == row[5]
-                        assert child2_library_name == row[6]
-                        assert child2_advantage_library_id == row[7]
-                        assert child2_token == row[8]
-                        assert "False" == row[9]
-                    last_index = index
-
-            os.remove(output_file_path)
-            assert last_index == 2
-            overdrive_api.get_advantage_accounts.assert_called_once()
